% Part of the oro.nifti package for R
% Distributed under the BSD license: see oro.nifti/COPYING
%
% $Id: $

\name{is.afni}
\alias{is.afni}
\title{is.afni}
\description{
  check object  
}
\usage{
is.afni(x)
}
\arguments{
  \item{x}{is an object to be checked.}
}
\details{
  Check whether object is of class \code{\linkS4class{afni}}  
}
\value{
  Logical indicating whether object is of class \code{\linkS4class{afni}}  
}
\references{
  AFNI\cr
  \url{http://afni.nimh.nih.gov/pub/dist/src/README.attributes}
}
\author{Karsten Tabelow \email{karsten.tabelow@wias-berlin.de}}
\seealso{
  \code{\linkS4class{afni}}
}
%\keyword{}
