% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_Operators.R
\name{anlz-nifti-ops}
\alias{anlz-nifti-ops}
\alias{Ops,anlz,anlz-method}
\alias{Ops,anlz,numeric-method}
\alias{Ops,numeric,anlz-method}
\alias{Ops,nifti,anlz-method}
\alias{Ops,anlz,nifti-method}
\title{Operations for Objects in the ANALYZE and NIfTI classes}
\usage{
\S4method{Ops}{anlz,anlz}(e1, e2)

\S4method{Ops}{anlz,numeric}(e1, e2)

\S4method{Ops}{numeric,anlz}(e1, e2)

\S4method{Ops}{nifti,anlz}(e1, e2)

\S4method{Ops}{anlz,nifti}(e1, e2)
}
\arguments{
\item{e1}{object}

\item{e2}{object}
}
\description{
Overloaded operators for anlz and nifti objects
}
\examples{

img01 <- anlz(array(1:64, c(4,4,4,1)), datatype=4)
img02 <- anlz(array(64:1, c(4,4,4,1)), datatype=4)
is.anlz(img01 + img02)
is.anlz(sqrt(2) * img01)
is.anlz(img02 / pi)

}
\author{
John Muschelli \email{muschellij2@gmail.com}
}
