% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/anlz_Operators.R
\docType{methods}
\name{anlz_niftiops}
\alias{Ops,anlz,nifti-method}
\alias{Ops,nifti,anlz-method}
\alias{anlz_niftiops}
\title{Operations for anlz and nifti Objects}
\usage{
\S4method{Ops}{nifti,anlz}(e1, e2)

\S4method{Ops}{anlz,nifti}(e1, e2)
}
\arguments{
\item{e1}{object}

\item{e2}{object}
}
\description{
Operations for anlz and nifti Objects
}

