% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/anlz_Operators.R
\docType{methods}
\name{anlzops}
\alias{Ops,anlz,anlz-method}
\alias{Ops,anlz,numeric-method}
\alias{Ops,numeric,anlz-method}
\alias{anlzops}
\title{Operations for anlz Objects}
\usage{
\S4method{Ops}{anlz,anlz}(e1, e2)

\S4method{Ops}{anlz,numeric}(e1, e2)

\S4method{Ops}{numeric,anlz}(e1, e2)
}
\arguments{
\item{e1}{object}

\item{e2}{object}
}
\description{
Operations for anlz Objects
}

