% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_anlz.R
\name{as.anlz}
\alias{as.anlz}
\title{as.anlz}
\usage{
as.anlz(from, value = NULL, verbose = FALSE)
}
\arguments{
\item{from}{is the object to be converted.}

\item{value}{is the \code{nifti} class object to use as a template for
various ANALYZE header information.}

\item{verbose}{is a logical variable (default = \code{FALSE}) that allows
text-based feedback during execution of the function.}
}
\value{
An object of class \code{anlz}.
}
\description{
Internal function that converts multidimensional arrays to ANALYZE class
objects.
}
\author{
Andrew Thornton \email{zeripath@users.sourceforge.net},\cr
Brandon Whitcher \email{bwhitcher@gmail.com}
}
