% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_nifti.R
\name{as.nifti}
\alias{as.nifti}
\title{as.nifti}
\usage{
as.nifti(from, value = NULL, verbose = FALSE)
}
\arguments{
\item{from}{is the object to be converted.}

\item{value}{is the \code{anlz} class object to use as a template for
various NIfTI header information.}

\item{verbose}{is a logical variable (default = \code{FALSE}) that allows
text-based feedback during execution of the function.}
}
\value{
An object of class \code{nifti}.
}
\description{
Internal function that converts multidimensional arrays to 
NIfTI class objects.
}
\author{
Andrew Thornton \email{zeripath@users.sourceforge.net}.\cr
Brandon Whitcher \email{bwhitcher@gmail.com}
}
