% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/aux_file.R
\docType{methods}
\name{aux.file}
\alias{aux.file}
\alias{aux.file-methods}
\title{Extract NIfTI 3D Image aux.file attribute}
\usage{
aux.file(object)
}
\arguments{
\item{object}{is an object of class \code{nifti}}

\item{value}{Value to assign to aux_file}
}
\description{
Methods that act on the ``aux_file'' in the NIfTI header.
}
\examples{
\dontrun{
url <- "http://nifti.nimh.nih.gov/nifti-1/data/avg152T1_RL_nifti.nii.gz"
urlfile <- file.path(system.file("nifti", package="oro.nifti"),
                    "mniRL.nii.gz")
download.file(url, urlfile, quiet=TRUE)
}
options("niftiAuditTrail"=FALSE)

urlfile <- file.path(system.file("nifti", package="oro.nifti"),
                     "mniRL.nii.gz")
mniRL <- readNIfTI(urlfile)
aux.file(mniRL)
aux.file(mniRL) <- "avg152T1_RL_nifti"
aux.file(mniRL)
}

