% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/bitpix.R
\docType{methods}
\name{bitpix}
\alias{bitpix}
\alias{bitpix,anlz-method}
\alias{bitpix,nifti-method}
\alias{bitpix-methods}
\alias{bitpix<-}
\alias{bitpix<-,anlz-method}
\alias{bitpix<-,nifti-method}
\title{Extract NIfTI 3D Image bitpix attribute}
\usage{
bitpix(object)

\S4method{bitpix}{nifti}(object)

\S4method{bitpix}{anlz}(object)

bitpix(object) <- value

\S4method{bitpix}{nifti}(object) <- value

\S4method{bitpix}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti}}

\item{value}{Value to assign to bitpix}
}
\description{
Methods that act on the ``bitpix'' in the NIfTI header.
}

