% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/cal.min.R
\docType{methods}
\name{cal.min}
\alias{cal.min}
\alias{cal.min,anlz-method}
\alias{cal.min,nifti-method}
\alias{cal.min-methods}
\alias{cal.min<-}
\alias{cal.min<-,anlz-method}
\alias{cal.min<-,nifti-method}
\title{Extract NIfTI 3D Image cal.min attribute}
\usage{
cal.min(object)

\S4method{cal.min}{nifti}(object)

\S4method{cal.min}{anlz}(object)

cal.min(object) <- value

\S4method{cal.min}{nifti}(object) <- value

\S4method{cal.min}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti}}

\item{value}{Value to assign to cal_min}
}
\description{
Methods that act on the ``cal_min'' in the NIfTI header.
}
\examples{
\dontrun{
url <- "http://nifti.nimh.nih.gov/nifti-1/data/avg152T1_LR_nifti.nii.gz"
urlfile <- file.path(system.file("nifti", package="oro.nifti"),
                     "mniLR.nii.gz")
download.file(url, urlfile, quiet=TRUE)
}
urlfile <- file.path(system.file("nifti", package="oro.nifti"),
                     "mniLR.nii.gz")
mniLR <- readNIfTI(urlfile)
cal.min(mniLR)
}

