% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/cal_max.R
\docType{methods}
\name{cal.max}
\alias{cal.max}
\alias{cal.max,anlz-method}
\alias{cal.max,nifti-method}
\alias{cal.max-methods}
\alias{cal.max<-}
\alias{cal.max<-,anlz-method}
\alias{cal.max<-,nifti-method}
\title{Extract Image cal.max attribute}
\usage{
cal.max(object)

\S4method{cal.max}{nifti}(object)

\S4method{cal.max}{anlz}(object)

cal.max(object) <- value

\S4method{cal.max}{nifti}(object) <- value

\S4method{cal.max}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to cal_max}
}
\description{
Methods that act on the ``cal_max'' in the NIfTI/ANALYZE header.
}
\examples{
\dontrun{
url <- "http://nifti.nimh.nih.gov/nifti-1/data/avg152T1_LR_nifti.nii.gz"
urlfile <- file.path(system.file("nifti", package="oro.nifti"),
                     "mniLR.nii.gz")
download.file(url, urlfile, quiet=TRUE)
}
urlfile <- file.path(system.file("nifti", package="oro.nifti"),
                     "mniLR.nii.gz")
mniLR <- readNIfTI(urlfile)
cal.max(mniLR)
}

