% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/cal_units.R
\docType{methods}
\name{cal_units}
\alias{cal_units}
\alias{cal_units,anlz-method}
\alias{cal_units,nifti-method}
\alias{cal_units-methods}
\alias{cal_units<-}
\alias{cal_units<-,anlz-method}
\alias{cal_units<-,nifti-method}
\title{Extract Image cal_units attribute}
\usage{
cal_units(object)

\S4method{cal_units}{nifti}(object)

\S4method{cal_units}{anlz}(object)

cal_units(object) <- value

\S4method{cal_units}{nifti}(object) <- value

\S4method{cal_units}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to cal_units}
}
\description{
Methods that act on the ``cal_units'' in the NIfTI/ANALYZE header.
}

