% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/compressed.R
\docType{methods}
\name{compressed-methods}
\alias{compressed}
\alias{compressed,anlz-method}
\alias{compressed,nifti-method}
\alias{compressed-methods}
\alias{compressed<-}
\alias{compressed<-,anlz-method}
\alias{compressed<-,nifti-method}
\title{Extract Image compressed attribute}
\usage{
compressed(object)

\S4method{compressed}{nifti}(object)

\S4method{compressed}{anlz}(object)

compressed(object) <- value

\S4method{compressed}{nifti}(object) <- value

\S4method{compressed}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to compressed}
}
\description{
Methods that act on the ``compressed'' in the NIfTI/ANALYZE header.
}

