% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/data_type.R
\docType{methods}
\name{data_type-methods}
\alias{data_type}
\alias{data_type,anlz-method}
\alias{data_type,nifti-method}
\alias{data_type-methods}
\alias{data_type<-}
\alias{data_type<-,anlz-method}
\alias{data_type<-,nifti-method}
\title{Extract Image data_type attribute}
\usage{
data_type(object)

\S4method{data_type}{nifti}(object)

\S4method{data_type}{anlz}(object)

data_type(object) <- value

\S4method{data_type}{nifti}(object) <- value

\S4method{data_type}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to data_type}
}
\description{
Methods that act on the ``data_type'' in the NIfTI/ANALYZE header.
}

