% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/datatype.R
\docType{methods}
\name{datatype}
\alias{datatype}
\alias{datatype,anlz-method}
\alias{datatype,nifti-method}
\alias{datatype-methods}
\alias{datatype<-}
\alias{datatype<-,anlz-method}
\alias{datatype<-,nifti-method}
\title{Extract NIfTI 3D Image datatype attribute}
\usage{
datatype(object)

\S4method{datatype}{nifti}(object)

\S4method{datatype}{anlz}(object)

datatype(object) <- value

\S4method{datatype}{nifti}(object) <- value

\S4method{datatype}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti}}

\item{value}{Value to assign to datatype}
}
\description{
Methods that act on the ``datatype'' in the NIfTI header.
}

