% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/db_name.R
\docType{methods}
\name{db_name-methods}
\alias{db_name}
\alias{db_name,anlz-method}
\alias{db_name,nifti-method}
\alias{db_name-methods}
\alias{db_name<-}
\alias{db_name<-,anlz-method}
\alias{db_name<-,nifti-method}
\title{Extract Image db_name attribute}
\usage{
db_name(object)

\S4method{db_name}{nifti}(object)

\S4method{db_name}{anlz}(object)

db_name(object) <- value

\S4method{db_name}{nifti}(object) <- value

\S4method{db_name}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to db_name}
}
\description{
Methods that act on the ``db_name'' in the NIfTI/ANALYZE header.
}

