% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/dim_.R
\docType{methods}
\name{dim_}
\alias{dim_}
\alias{dim_,anlz-method}
\alias{dim_,nifti-method}
\alias{dim_-methods}
\alias{dim_<-}
\alias{dim_<-,anlz-method}
\alias{dim_<-,nifti-method}
\title{Extract NIfTI 3D Image dim_ attribute}
\usage{
dim_(object)

\S4method{dim_}{nifti}(object)

\S4method{dim_}{anlz}(object)

dim_(object) <- value

\S4method{dim_}{nifti}(object) <- value

\S4method{dim_}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti}}

\item{value}{Value to assign to dim_}
}
\description{
Methods that act on the ``dim_'' in the NIfTI header.
}

