% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dim_info.R
\docType{methods}
\name{dim_info-methods}
\alias{dim_info}
\alias{dim_info,anlz-method}
\alias{dim_info,nifti-method}
\alias{dim_info-methods}
\alias{dim_info<-}
\alias{dim_info<-,anlz-method}
\alias{dim_info<-,nifti-method}
\title{Extract Image dim_info attribute}
\usage{
dim_info(object)

\S4method{dim_info}{nifti}(object)

\S4method{dim_info}{anlz}(object)

dim_info(object) <- value

\S4method{dim_info}{nifti}(object) <- value

\S4method{dim_info}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to dim_info}
}
\description{
Methods that act on the ``dim_info'' in the NIfTI/ANALYZE header.
}

