% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/dim_un0.R
\docType{methods}
\name{dim_un0}
\alias{dim_un0}
\alias{dim_un0,anlz-method}
\alias{dim_un0,nifti-method}
\alias{dim_un0-methods}
\alias{dim_un0<-}
\alias{dim_un0<-,anlz-method}
\alias{dim_un0<-,nifti-method}
\title{Extract Image dim_un0 attribute}
\usage{
dim_un0(object)

\S4method{dim_un0}{nifti}(object)

\S4method{dim_un0}{anlz}(object)

dim_un0(object) <- value

\S4method{dim_un0}{nifti}(object) <- value

\S4method{dim_un0}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to dim_un0}
}
\description{
Methods that act on the ``dim_un0'' in the NIfTI/ANALYZE header.
}

