% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/drop_img_dim.R
\name{drop_img_dim}
\alias{drop_img_dim}
\title{Drop Image Dimension}
\usage{
drop_img_dim(img, warn = TRUE)
}
\arguments{
\item{img}{nifti object}

\item{warn}{output warning if dimensions are under 3}
}
\value{
Object of class nifti
}
\description{
Drops a dimension of an image that has one-dimension and
sets respective values to 0 in pixdim or 1 in dim
}
\examples{
\dontrun{
## 27 scans of Colin Holmes (MNI) brain co-registered and averaged
## NIfTI two-file format
URL <- "http://imaging.mrc-cbu.cam.ac.uk/downloads/Colin/colin_1mm.tgz"
urlfile <- file.path(tempdir(), "colin_1mm.tgz")
download.file(URL, dest=urlfile, quiet=TRUE)
untar(urlfile, exdir=tempdir())
colin <- readNIfTI(file.path(tempdir(), "colin_1mm"))
dim(colin)
dim_(colin)
pixdim(colin)
# this will error
writeNIfTI(colin, filename = tempfile())
colin = drop_img_dim(colin)
writeNIfTI(colin, filename = tempfile())
}
}

