% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/exp_date.R
\docType{methods}
\name{exp_date}
\alias{exp_date}
\alias{exp_date,anlz-method}
\alias{exp_date,nifti-method}
\alias{exp_date-methods}
\alias{exp_date<-}
\alias{exp_date<-,anlz-method}
\alias{exp_date<-,nifti-method}
\title{Extract Image exp_date attribute}
\usage{
exp_date(object)

\S4method{exp_date}{nifti}(object)

\S4method{exp_date}{anlz}(object)

exp_date(object) <- value

\S4method{exp_date}{nifti}(object) <- value

\S4method{exp_date}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to exp_date}
}
\description{
Methods that act on the ``exp_date'' in the NIfTI/ANALYZE header.
}

