% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/exp_time.R
\docType{methods}
\name{exp_time-methods}
\alias{exp_time}
\alias{exp_time,anlz-method}
\alias{exp_time,nifti-method}
\alias{exp_time-methods}
\alias{exp_time<-}
\alias{exp_time<-,anlz-method}
\alias{exp_time<-,nifti-method}
\title{Extract Image exp_time attribute}
\usage{
exp_time(object)

\S4method{exp_time}{nifti}(object)

\S4method{exp_time}{anlz}(object)

exp_time(object) <- value

\S4method{exp_time}{nifti}(object) <- value

\S4method{exp_time}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to exp_time}
}
\description{
Methods that act on the ``exp_time'' in the NIfTI/ANALYZE header.
}

