% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/extender.R
\docType{methods}
\name{extender-methods}
\alias{extender}
\alias{extender,anlz-method}
\alias{extender,nifti-method}
\alias{extender-methods}
\alias{extender<-}
\alias{extender<-,anlz-method}
\alias{extender<-,nifti-method}
\title{Extract Image extender attribute}
\usage{
extender(object)

\S4method{extender}{nifti}(object)

\S4method{extender}{anlz}(object)

extender(object) <- value

\S4method{extender}{nifti}(object) <- value

\S4method{extender}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to extender}
}
\description{
Methods that act on the ``extender'' in the NIfTI/ANALYZE header.
}

