% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extents.R
\docType{methods}
\name{extents-methods}
\alias{extents}
\alias{extents,anlz-method}
\alias{extents,nifti-method}
\alias{extents-methods}
\alias{extents<-}
\alias{extents<-,anlz-method}
\alias{extents<-,nifti-method}
\title{Extract Image extents attribute}
\usage{
extents(object)

\S4method{extents}{nifti}(object)

\S4method{extents}{anlz}(object)

extents(object) <- value

\S4method{extents}{nifti}(object) <- value

\S4method{extents}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to extents}
}
\description{
Methods that act on the ``extents'' in the NIfTI/ANALYZE header.
}

