% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/field_skip.R
\docType{methods}
\name{field_skip}
\alias{field_skip}
\alias{field_skip,anlz-method}
\alias{field_skip,nifti-method}
\alias{field_skip-methods}
\alias{field_skip<-}
\alias{field_skip<-,anlz-method}
\alias{field_skip<-,nifti-method}
\title{Extract Image field_skip attribute}
\usage{
field_skip(object)

\S4method{field_skip}{nifti}(object)

\S4method{field_skip}{anlz}(object)

field_skip(object) <- value

\S4method{field_skip}{nifti}(object) <- value

\S4method{field_skip}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to field_skip}
}
\description{
Methods that act on the ``field_skip'' in the NIfTI/ANALYZE header.
}

