% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/funused1.R
\docType{methods}
\name{funused1}
\alias{funused1}
\alias{funused1,anlz-method}
\alias{funused1,nifti-method}
\alias{funused1-methods}
\alias{funused1<-}
\alias{funused1<-,anlz-method}
\alias{funused1<-,nifti-method}
\title{Extract Image funused1 attribute}
\usage{
funused1(object)

\S4method{funused1}{nifti}(object)

\S4method{funused1}{anlz}(object)

funused1(object) <- value

\S4method{funused1}{nifti}(object) <- value

\S4method{funused1}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to funused1}
}
\description{
Methods that act on the ``funused1'' in the NIfTI/ANALYZE header.
}

