% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/funused2.R
\docType{methods}
\name{funused2}
\alias{funused2}
\alias{funused2,anlz-method}
\alias{funused2,nifti-method}
\alias{funused2-methods}
\alias{funused2<-}
\alias{funused2<-,anlz-method}
\alias{funused2<-,nifti-method}
\title{Extract Image funused2 attribute}
\usage{
funused2(object)

\S4method{funused2}{nifti}(object)

\S4method{funused2}{anlz}(object)

funused2(object) <- value

\S4method{funused2}{nifti}(object) <- value

\S4method{funused2}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to funused2}
}
\description{
Methods that act on the ``funused2'' in the NIfTI/ANALYZE header.
}

