% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/funused3.R
\docType{methods}
\name{funused3}
\alias{funused3}
\alias{funused3,anlz-method}
\alias{funused3,nifti-method}
\alias{funused3-methods}
\alias{funused3<-}
\alias{funused3<-,anlz-method}
\alias{funused3<-,nifti-method}
\title{Extract Image funused3 attribute}
\usage{
funused3(object)

\S4method{funused3}{nifti}(object)

\S4method{funused3}{anlz}(object)

funused3(object) <- value

\S4method{funused3}{nifti}(object) <- value

\S4method{funused3}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to funused3}
}
\description{
Methods that act on the ``funused3'' in the NIfTI/ANALYZE header.
}

