% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generated.R
\docType{methods}
\name{generated-methods}
\alias{generated}
\alias{generated,anlz-method}
\alias{generated,nifti-method}
\alias{generated-methods}
\alias{generated<-}
\alias{generated<-,anlz-method}
\alias{generated<-,nifti-method}
\title{Extract Image generated attribute}
\usage{
generated(object)

\S4method{generated}{nifti}(object)

\S4method{generated}{anlz}(object)

generated(object) <- value

\S4method{generated}{nifti}(object) <- value

\S4method{generated}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to generated}
}
\description{
Methods that act on the ``generated'' in the NIfTI/ANALYZE header.
}

