% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/glmax.R
\docType{methods}
\name{glmax-methods}
\alias{glmax}
\alias{glmax,anlz-method}
\alias{glmax,nifti-method}
\alias{glmax-methods}
\alias{glmax<-}
\alias{glmax<-,anlz-method}
\alias{glmax<-,nifti-method}
\title{Extract Image glmax attribute}
\usage{
glmax(object)

\S4method{glmax}{nifti}(object)

\S4method{glmax}{anlz}(object)

glmax(object) <- value

\S4method{glmax}{nifti}(object) <- value

\S4method{glmax}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to glmax}
}
\description{
Methods that act on the ``glmax'' in the NIfTI/ANALYZE header.
}

