% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/glmin.R
\docType{methods}
\name{glmin-methods}
\alias{glmin}
\alias{glmin,anlz-method}
\alias{glmin,nifti-method}
\alias{glmin-methods}
\alias{glmin<-}
\alias{glmin<-,anlz-method}
\alias{glmin<-,nifti-method}
\title{Extract Image glmin attribute}
\usage{
glmin(object)

\S4method{glmin}{nifti}(object)

\S4method{glmin}{anlz}(object)

glmin(object) <- value

\S4method{glmin}{nifti}(object) <- value

\S4method{glmin}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to glmin}
}
\description{
Methods that act on the ``glmin'' in the NIfTI/ANALYZE header.
}

