% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hist_un0.R
\docType{methods}
\name{hist_un0-methods}
\alias{hist_un0}
\alias{hist_un0,anlz-method}
\alias{hist_un0,nifti-method}
\alias{hist_un0-methods}
\alias{hist_un0<-}
\alias{hist_un0<-,anlz-method}
\alias{hist_un0<-,nifti-method}
\title{Extract Image hist_un0 attribute}
\usage{
hist_un0(object)

\S4method{hist_un0}{nifti}(object)

\S4method{hist_un0}{anlz}(object)

hist_un0(object) <- value

\S4method{hist_un0}{nifti}(object) <- value

\S4method{hist_un0}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to hist_un0}
}
\description{
Methods that act on the ``hist_un0'' in the NIfTI/ANALYZE header.
}

