% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/hkey_un0.R
\docType{methods}
\name{hkey_un0}
\alias{hkey_un0}
\alias{hkey_un0,anlz-method}
\alias{hkey_un0,nifti-method}
\alias{hkey_un0-methods}
\alias{hkey_un0<-}
\alias{hkey_un0<-,anlz-method}
\alias{hkey_un0<-,nifti-method}
\title{Extract Image hkey_un0 attribute}
\usage{
hkey_un0(object)

\S4method{hkey_un0}{nifti}(object)

\S4method{hkey_un0}{anlz}(object)

hkey_un0(object) <- value

\S4method{hkey_un0}{nifti}(object) <- value

\S4method{hkey_un0}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to hkey_un0}
}
\description{
Methods that act on the ``hkey_un0'' in the NIfTI/ANALYZE header.
}

