% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_data.R
\docType{methods}
\name{img_data-methods}
\alias{img_data-methods}
\alias{img_data-methods,}
\alias{img_data}
\alias{img_data,nifti-method}
\alias{img_data,anlz-method}
\alias{img_data,character-method}
\alias{img_data,ANY-method}
\alias{img_data<-}
\alias{img_data<-,nifti-method}
\alias{img_data<-,anlz-method}
\title{Extract Image Attribute \code{.Data}}
\usage{
img_data(object)

\S4method{img_data}{nifti}(object)

\S4method{img_data}{anlz}(object)

\S4method{img_data}{character}(object)

\S4method{img_data}{ANY}(object)

img_data(object) <- value

\S4method{img_data}{nifti}(object) <- value

\S4method{img_data}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}.}

\item{value}{is the value to assign to the \code{.Data} field.}
}
\description{
Methods that act on the \code{.Data} field in the
NIfTI/ANALYZE header.
}
\details{
See documentation on the ANALYZE and/or NIfTI data standards for
more details.
}
\references{
ANALYZE 7.5\cr
\url{http://eeg.sourceforge.net/ANALYZE75.pdf}\cr
NIfTI-1\cr
\url{http://nifti.nimh.nih.gov/}
}
\author{
John Muschelli \email{muschellij2@gmail.com},\cr
Brandon Whitcher \email{bwhitcher@gmail.com}
}
