<<<<<<< HEAD
% Part of the oro.nifti package for R
% Distributed under the BSD license: see oro.nifti/COPYING
%
% $Id: $

=======
% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/transform.R
>>>>>>> moro.nifti-master
\name{integerTranslation}
\alias{integerTranslation}
\alias{invertIntegerTranslation}
\title{integerTranslation}
<<<<<<< HEAD
\description{
  ...
}
\usage{
integerTranslation(nim, data, verbose = FALSE)
invertIntegerTranslation(nim, verbose = FALSE)
}
\arguments{
  \item{nim}{is an object of class \code{nifti}.}
  \item{data}{is ...}
  \item{verbose}{is a logical variable (default = \code{FALSE}) that
    allows text-based feedback during execution of the function.}
}
\details{
  ...
}
\value{
  ...
}
%\references{}
%\seealso{}
%\examples{}
\author{Andrew Thornton \email{zeripath@users.sourceforge.net}}
%\keyword{}
=======
\usage{
integerTranslation(nim, data, verbose = FALSE)

invertIntegerTranslation(nim, verbose = FALSE)
}
\arguments{
\item{nim}{is an object of class \code{nifti}.}

\item{data}{is ...}

\item{verbose}{is a logical variable (default = \code{FALSE}) that allows
text-based feedback during execution of the function.}
}
\value{
...
}
\description{
...
}
\details{
...
}
\author{
Andrew Thornton \email{zeripath@users.sourceforge.net}
}

>>>>>>> moro.nifti-master
