% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/intent_name.R
\docType{methods}
\name{intent_name-methods}
\alias{intent_name}
\alias{intent_name,anlz-method}
\alias{intent_name,nifti-method}
\alias{intent_name-methods}
\alias{intent_name<-}
\alias{intent_name<-,anlz-method}
\alias{intent_name<-,nifti-method}
\title{Extract Image intent_name attribute}
\usage{
intent_name(object)

\S4method{intent_name}{nifti}(object)

\S4method{intent_name}{anlz}(object)

intent_name(object) <- value

\S4method{intent_name}{nifti}(object) <- value

\S4method{intent_name}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to intent_name}
}
\description{
Methods that act on the ``intent_name'' in the NIfTI/ANALYZE header.
}

