% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/intent_p1.R
\docType{methods}
\name{intent_p1-methods}
\alias{intent_p1}
\alias{intent_p1,anlz-method}
\alias{intent_p1,nifti-method}
\alias{intent_p1-methods}
\alias{intent_p1<-}
\alias{intent_p1<-,anlz-method}
\alias{intent_p1<-,nifti-method}
\title{Extract Image intent_p1 attribute}
\usage{
intent_p1(object)

\S4method{intent_p1}{nifti}(object)

\S4method{intent_p1}{anlz}(object)

intent_p1(object) <- value

\S4method{intent_p1}{nifti}(object) <- value

\S4method{intent_p1}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to intent_p1}
}
\description{
Methods that act on the ``intent_p1'' in the NIfTI/ANALYZE header.
}

