% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/intent_p2.R
\docType{methods}
\name{intent_p2-methods}
\alias{intent_p2}
\alias{intent_p2,anlz-method}
\alias{intent_p2,nifti-method}
\alias{intent_p2-methods}
\alias{intent_p2<-}
\alias{intent_p2<-,anlz-method}
\alias{intent_p2<-,nifti-method}
\title{Extract Image intent_p2 attribute}
\usage{
intent_p2(object)

\S4method{intent_p2}{nifti}(object)

\S4method{intent_p2}{anlz}(object)

intent_p2(object) <- value

\S4method{intent_p2}{nifti}(object) <- value

\S4method{intent_p2}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to intent_p2}
}
\description{
Methods that act on the ``intent_p2'' in the NIfTI/ANALYZE header.
}

