% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/intent_p3.R
\docType{methods}
\name{intent_p3}
\alias{intent_p3}
\alias{intent_p3,anlz-method}
\alias{intent_p3,nifti-method}
\alias{intent_p3-methods}
\alias{intent_p3<-}
\alias{intent_p3<-,anlz-method}
\alias{intent_p3<-,nifti-method}
\title{Extract Image intent_p3 attribute}
\usage{
intent_p3(object)

\S4method{intent_p3}{nifti}(object)

\S4method{intent_p3}{anlz}(object)

intent_p3(object) <- value

\S4method{intent_p3}{nifti}(object) <- value

\S4method{intent_p3}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to intent_p3}
}
\description{
Methods that act on the ``intent_p3'' in the NIfTI/ANALYZE header.
}

