% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.afni}
\alias{is.afni}
\title{check object}
\usage{
is.afni(x)
}
\arguments{
\item{x}{is an object to be checked.}
}
\value{
Logical indicating whether object is of class
\code{\linkS4class{afni}}.
}
\description{
Check whether object is of class \code{\linkS4class{afni}}.
}
\references{
AFNI\cr
\url{http://afni.nimh.nih.gov/pub/dist/src/README.attributes}
}
\seealso{
\code{\linkS4class{afni}}
}
\author{
Karsten Tabelow \email{karsten.tabelow@wias-berlin.de}
}
