% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.anlz}
\alias{is.anlz}
\title{check object}
\usage{
is.anlz(x)
}
\arguments{
\item{x}{is an object to be checked.}
}
\value{
Logical indicating whether object is of class
\code{\linkS4class{anlz}}.
}
\description{
Check whether object is of class \code{\linkS4class{anlz}}.
}
\references{
ANALYZE 7.5\cr \url{http://eeg.sourceforge.net/ANALYZE75.pdf}
}
\seealso{
\code{\linkS4class{anlz}}
}
\author{
Karsten Tabelow \email{karsten.tabelow@wias-berlin.de}
}
