% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/niftiS4.R
\name{is.nifti}
\alias{is.nifti}
\title{check object}
\usage{
is.nifti(x)
}
\arguments{
\item{x}{is an object to be checked.}
}
\value{
Logical indicating whether object is of class
\code{\linkS4class{nifti}}
}
\description{
Check whether object is of class \code{\linkS4class{nifti}}
}
\author{
Karsten Tabelow \email{karsten.tabelow@wias-berlin.de}
}
\references{
ANALYZE 7.5\cr \url{http://www.mayo.edu/bir/PDF/ANALYZE75.pdf}
}
\seealso{
\code{\linkS4class{nifti}}
}

