% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/magic.R
\docType{methods}
\name{magic}
\alias{magic}
\alias{magic,anlz-method}
\alias{magic,nifti-method}
\alias{magic-methods}
\alias{magic<-}
\alias{magic<-,anlz-method}
\alias{magic<-,nifti-method}
\title{Extract NIfTI 3D Image magic attribute}
\usage{
magic(object)

\S4method{magic}{nifti}(object)

\S4method{magic}{anlz}(object)

magic(object) <- value

\S4method{magic}{nifti}(object) <- value

\S4method{magic}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti}}

\item{value}{Value to assign to magic}
}
\description{
Methods that act on the ``magic'' in the NIfTI header.
}

