% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niftiS4.R
\docType{class}
\name{niftiExtension-class}
\alias{niftiExtension-class}
\title{Class "niftiExtension"}
\description{
An extension of the NIfTI class that allows \dQuote{extensions}
that conform to the NIfTI data standard.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("niftiExtension", data, dim, dimnames, ...)}.
}

\examples{
showClass("niftiExtension")
}
\references{
NIfTI-1\cr \url{http://nifti.nimh.nih.gov/}
}
\seealso{
\code{\linkS4class{nifti}}, \code{\linkS4class{niftiAuditTrail}}
}
\author{
Andrew Thornton \email{zeripath@users.sourceforge.net}
}
\keyword{classes}
