% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Operators.R
\docType{methods}
\name{niftiops}
\alias{Ops,nifti,nifti-method}
\alias{Ops,nifti,numeric-method}
\alias{Ops,numeric,nifti-method}
\alias{niftiops}
\title{Operations for NIfTI Objects}
\usage{
\S4method{Ops}{nifti,nifti}(e1, e2)

\S4method{Ops}{nifti,numeric}(e1, e2)

\S4method{Ops}{numeric,nifti}(e1, e2)
}
\arguments{
\item{e1}{is an object of class \code{nifti}.}

\item{e2}{is an object of class \code{nifti}.}
}
\description{
Operations for NIfTI Objects
}
\examples{
img04 <- nifti()
img05 <- nifti()
img06 <- img04 + img05
}
\author{
John Muschellli \email{muschellij2@gmail.com}
}

