% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/omax.R
\docType{methods}
\name{omax-methods}
\alias{omax}
\alias{omax,anlz-method}
\alias{omax,nifti-method}
\alias{omax-methods}
\alias{omax<-}
\alias{omax<-,anlz-method}
\alias{omax<-,nifti-method}
\title{Extract Image omax attribute}
\usage{
omax(object)

\S4method{omax}{nifti}(object)

\S4method{omax}{anlz}(object)

omax(object) <- value

\S4method{omax}{nifti}(object) <- value

\S4method{omax}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to omax}
}
\description{
Methods that act on the ``omax'' in the NIfTI/ANALYZE header.
}

