% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/omin.R
\docType{methods}
\name{omin-methods}
\alias{omin}
\alias{omin,anlz-method}
\alias{omin,nifti-method}
\alias{omin-methods}
\alias{omin<-}
\alias{omin<-,anlz-method}
\alias{omin<-,nifti-method}
\title{Extract Image omin attribute}
\usage{
omin(object)

\S4method{omin}{nifti}(object)

\S4method{omin}{anlz}(object)

omin(object) <- value

\S4method{omin}{nifti}(object) <- value

\S4method{omin}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to omin}
}
\description{
Methods that act on the ``omin'' in the NIfTI/ANALYZE header.
}

