% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/orient.R
\docType{methods}
\name{orient-methods}
\alias{orient}
\alias{orient,anlz-method}
\alias{orient,nifti-method}
\alias{orient-methods}
\alias{orient<-}
\alias{orient<-,anlz-method}
\alias{orient<-,nifti-method}
\title{Extract Image orient attribute}
\usage{
orient(object)

\S4method{orient}{nifti}(object)

\S4method{orient}{anlz}(object)

orient(object) <- value

\S4method{orient}{nifti}(object) <- value

\S4method{orient}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to orient}
}
\description{
Methods that act on the ``orient'' in the NIfTI/ANALYZE header.
}

