% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/niftiS4.R, R/qform.R, R/sform.R
\docType{methods}
\name{orientation-methods}
\alias{orientation-methods}
\alias{qform}
\alias{qform,character}
\alias{qform,character-method}
\alias{qform,nifti-method}
\alias{qform-methods}
\alias{sform}
\alias{sform,character}
\alias{sform,character-method}
\alias{sform,nifti-method}
\alias{sform-methods}
\title{Extract NIfTI 3D Image Orientation}
\usage{
sform(object)

\S4method{sform}{nifti}(object)

qform(object)

\S4method{qform}{nifti}(object)

\S4method{qform}{character}(object)

\S4method{sform}{character}(object)
}
\arguments{
\item{object}{is an object of class \code{nifti}.}

\item{object}{is a \code{nifti} object}
}
\description{
Methods that act on the \dQuote{qform} and \dQuote{sform} information in the
NIfTI header.

Extract NIfTI 3D Image Orientation

Extract NIfTI 3D Image Orientation
}
\section{Methods}{
 \describe{ \item{object = "nifti"}{Extract or replace
NIfTI description.} }
}
\examples{
\dontrun{
url <- "http://nifti.nimh.nih.gov/nifti-1/data/avg152T1_LR_nifti.nii.gz"
urlfile <- file.path(system.file("nifti", package="oro.nifti"),
                     "mniLR.nii.gz")
download.file(url, urlfile, quiet=TRUE)
}
urlfile <- file.path(system.file("nifti", package="oro.nifti"),
                     "mniLR.nii.gz")
mniLR <- readNIfTI(urlfile)
sform(mniLR)
}
\author{
Brandon Whitcher \email{bwhitcher@gmail.com}
}
\keyword{methods}

