% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/origin.R
\docType{methods}
\name{origin-methods}
\alias{origin}
\alias{origin,anlz-method}
\alias{origin,nifti-method}
\alias{origin-methods}
\alias{origin<-}
\alias{origin<-,anlz-method}
\alias{origin<-,nifti-method}
\title{Extract Image origin attribute}
\usage{
origin(object)

\S4method{origin}{nifti}(object)

\S4method{origin}{anlz}(object)

origin(object) <- value

\S4method{origin}{nifti}(object) <- value

\S4method{origin}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to origin}
}
\description{
Methods that act on the ``origin'' in the NIfTI/ANALYZE header.
}

