% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/patient_id.R
\docType{methods}
\name{patient_id-methods}
\alias{patient_id}
\alias{patient_id,anlz-method}
\alias{patient_id,nifti-method}
\alias{patient_id-methods}
\alias{patient_id<-}
\alias{patient_id<-,anlz-method}
\alias{patient_id<-,nifti-method}
\title{Extract Image patient_id attribute}
\usage{
patient_id(object)

\S4method{patient_id}{nifti}(object)

\S4method{patient_id}{anlz}(object)

patient_id(object) <- value

\S4method{patient_id}{nifti}(object) <- value

\S4method{patient_id}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to patient_id}
}
\description{
Methods that act on the ``patient_id'' in the NIfTI/ANALYZE header.
}

