% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qform_code.R
\docType{methods}
\name{qform_code-methods}
\alias{qform_code}
\alias{qform_code,anlz-method}
\alias{qform_code,nifti-method}
\alias{qform_code-methods}
\alias{qform_code<-}
\alias{qform_code<-,anlz-method}
\alias{qform_code<-,nifti-method}
\title{Extract Image qform_code attribute}
\usage{
qform_code(object)

\S4method{qform_code}{nifti}(object)

\S4method{qform_code}{anlz}(object)

qform_code(object) <- value

\S4method{qform_code}{nifti}(object) <- value

\S4method{qform_code}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to qform_code}
}
\description{
Methods that act on the ``qform_code'' in the NIfTI/ANALYZE header.
}

