% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/qoffset_x.R
\docType{methods}
\name{qoffset_x-methods}
\alias{qoffset_x}
\alias{qoffset_x,anlz-method}
\alias{qoffset_x,nifti-method}
\alias{qoffset_x-methods}
\alias{qoffset_x<-}
\alias{qoffset_x<-,anlz-method}
\alias{qoffset_x<-,nifti-method}
\title{Extract Image qoffset_x attribute}
\usage{
qoffset_x(object)

\S4method{qoffset_x}{nifti}(object)

\S4method{qoffset_x}{anlz}(object)

qoffset_x(object) <- value

\S4method{qoffset_x}{nifti}(object) <- value

\S4method{qoffset_x}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to qoffset_x}
}
\description{
Methods that act on the ``qoffset_x'' in the NIfTI/ANALYZE header.
}

