% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/qoffset_y.R
\docType{methods}
\name{qoffset_y}
\alias{qoffset_y}
\alias{qoffset_y,anlz-method}
\alias{qoffset_y,nifti-method}
\alias{qoffset_y-methods}
\alias{qoffset_y<-}
\alias{qoffset_y<-,anlz-method}
\alias{qoffset_y<-,nifti-method}
\title{Extract NIfTI 3D Image qoffset_y attribute}
\usage{
qoffset_y(object)

\S4method{qoffset_y}{nifti}(object)

\S4method{qoffset_y}{anlz}(object)

qoffset_y(object) <- value

\S4method{qoffset_y}{nifti}(object) <- value

\S4method{qoffset_y}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti}}

\item{value}{Value to assign to qoffset_y}
}
\description{
Methods that act on the ``qoffset_y'' in the NIfTI header.
}

