% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/qoffset_z.R
\docType{methods}
\name{qoffset_z}
\alias{qoffset_z}
\alias{qoffset_z,anlz-method}
\alias{qoffset_z,nifti-method}
\alias{qoffset_z-methods}
\alias{qoffset_z<-}
\alias{qoffset_z<-,anlz-method}
\alias{qoffset_z<-,nifti-method}
\title{Extract Image qoffset_z attribute}
\usage{
qoffset_z(object)

\S4method{qoffset_z}{nifti}(object)

\S4method{qoffset_z}{anlz}(object)

qoffset_z(object) <- value

\S4method{qoffset_z}{nifti}(object) <- value

\S4method{qoffset_z}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to qoffset_z}
}
\description{
Methods that act on the ``qoffset_z'' in the NIfTI/ANALYZE header.
}

