% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/quatern_b.R
\docType{methods}
\name{quatern_b}
\alias{quatern_b}
\alias{quatern_b,anlz-method}
\alias{quatern_b,nifti-method}
\alias{quatern_b-methods}
\alias{quatern_b<-}
\alias{quatern_b<-,anlz-method}
\alias{quatern_b<-,nifti-method}
\title{Extract NIfTI 3D Image quatern_b attribute}
\usage{
quatern_b(object)

\S4method{quatern_b}{nifti}(object)

\S4method{quatern_b}{anlz}(object)

quatern_b(object) <- value

\S4method{quatern_b}{nifti}(object) <- value

\S4method{quatern_b}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti}}

\item{value}{Value to assign to quatern_b}
}
\description{
Methods that act on the ``quatern_b'' in the NIfTI header.
}

