% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/quatern_c.R
\docType{methods}
\name{quatern_c}
\alias{quatern_c}
\alias{quatern_c,anlz-method}
\alias{quatern_c,nifti-method}
\alias{quatern_c-methods}
\alias{quatern_c<-}
\alias{quatern_c<-,anlz-method}
\alias{quatern_c<-,nifti-method}
\title{Extract NIfTI 3D Image quatern_c attribute}
\usage{
quatern_c(object)

\S4method{quatern_c}{nifti}(object)

\S4method{quatern_c}{anlz}(object)

quatern_c(object) <- value

\S4method{quatern_c}{nifti}(object) <- value

\S4method{quatern_c}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti}}

\item{value}{Value to assign to quatern_c}
}
\description{
Methods that act on the ``quatern_c'' in the NIfTI header.
}

