% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/quatern_d.R
\docType{methods}
\name{quatern_d}
\alias{quatern_d}
\alias{quatern_d,anlz-method}
\alias{quatern_d,nifti-method}
\alias{quatern_d-methods}
\alias{quatern_d<-}
\alias{quatern_d<-,anlz-method}
\alias{quatern_d<-,nifti-method}
\title{Extract Image quatern_d attribute}
\usage{
quatern_d(object)

\S4method{quatern_d}{nifti}(object)

\S4method{quatern_d}{anlz}(object)

quatern_d(object) <- value

\S4method{quatern_d}{nifti}(object) <- value

\S4method{quatern_d}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to quatern_d}
}
\description{
Methods that act on the ``quatern_d'' in the NIfTI/ANALYZE header.
}

