% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/regular.R
\docType{methods}
\name{regular}
\alias{regular}
\alias{regular,anlz-method}
\alias{regular,nifti-method}
\alias{regular-methods}
\alias{regular<-}
\alias{regular<-,anlz-method}
\alias{regular<-,nifti-method}
\title{Extract Image regular attribute}
\usage{
regular(object)

\S4method{regular}{nifti}(object)

\S4method{regular}{anlz}(object)

regular(object) <- value

\S4method{regular}{nifti}(object) <- value

\S4method{regular}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to regular}
}
\description{
Methods that act on the ``regular'' in the NIfTI/ANALYZE header.
}

