<<<<<<< HEAD
\name{rmniigz}
\alias{rmniigz}
\alias{rmnii}
\alias{rmgz}
\alias{rmhdrgz}
\alias{rmhdr}
\alias{rmimggz}
\alias{rmimg}
\title{Remove File Extensions Around the NIfTI/ANALYZE Formats}
\description{
  Simple function(s) that remove file extensions commonly found when
  using NIfTI-1 or ANALYZE format files.
}
\usage{
rmniigz(x)
rmnii(x)
rmgz(x)
rmhdrgz(x)
rmhdr(x)
rmimggz(x)
rmimg(x)
}
\arguments{
  \item{x}{is the file name.}
}
\value{
  The file name without offending suffix.
}
%\details{}
%\references{}
%\seealso{}
%\examples{}
\author{Brandon Whitcher \email{bwhitcher@gmail.com}}
\keyword{Misc}
=======
% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/remove.R
\name{rmniigz}
\alias{rmgz}
\alias{rmhdr}
\alias{rmhdrgz}
\alias{rmimg}
\alias{rmimggz}
\alias{rmnii}
\alias{rmniigz}
\title{Remove File Extensions Around the NIfTI/ANALYZE Formats}
\usage{
rmniigz(x)

rmnii(x)

rmgz(x)

rmhdrgz(x)

rmhdr(x)

rmimggz(x)

rmimg(x)
}
\arguments{
\item{x}{is the file name.}
}
\value{
The file name without offending suffix.
}
\description{
Simple function(s) that remove file extensions commonly found when using
NIfTI-1 or ANALYZE format files.
}
\author{
Brandon Whitcher \email{bwhitcher@gmail.com}
}
\keyword{Misc}

>>>>>>> moro.nifti-master
