% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/reoriented.R
\docType{methods}
\name{reoriented-methods}
\alias{reoriented}
\alias{reoriented,anlz-method}
\alias{reoriented,nifti-method}
\alias{reoriented-methods}
\alias{reoriented<-}
\alias{reoriented<-,anlz-method}
\alias{reoriented<-,nifti-method}
\title{Extract Image reoriented attribute}
\usage{
reoriented(object)

\S4method{reoriented}{nifti}(object)

\S4method{reoriented}{anlz}(object)

reoriented(object) <- value

\S4method{reoriented}{nifti}(object) <- value

\S4method{reoriented}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to reoriented}
}
\description{
Methods that act on the ``reoriented'' in the NIfTI/ANALYZE header.
}

