% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scannum.R
\docType{methods}
\name{scannum-methods}
\alias{scannum}
\alias{scannum,anlz-method}
\alias{scannum,nifti-method}
\alias{scannum-methods}
\alias{scannum<-}
\alias{scannum<-,anlz-method}
\alias{scannum<-,nifti-method}
\title{Extract Image scannum attribute}
\usage{
scannum(object)

\S4method{scannum}{nifti}(object)

\S4method{scannum}{anlz}(object)

scannum(object) <- value

\S4method{scannum}{nifti}(object) <- value

\S4method{scannum}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to scannum}
}
\description{
Methods that act on the ``scannum'' in the NIfTI/ANALYZE header.
}

