% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/scl_inter.R
\docType{methods}
\name{scl_inter}
\alias{scl_inter}
\alias{scl_inter,anlz-method}
\alias{scl_inter,nifti-method}
\alias{scl_inter-methods}
\alias{scl_inter<-}
\alias{scl_inter<-,anlz-method}
\alias{scl_inter<-,nifti-method}
\title{Extract Image scl_inter attribute}
\usage{
scl_inter(object)

\S4method{scl_inter}{nifti}(object)

\S4method{scl_inter}{anlz}(object)

scl_inter(object) <- value

\S4method{scl_inter}{nifti}(object) <- value

\S4method{scl_inter}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to scl_inter}
}
\description{
Methods that act on the ``scl_inter'' in the NIfTI/ANALYZE header.
}

