% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/scl_slope.R
\docType{methods}
\name{scl_slope-methods}
\alias{scl_slope}
\alias{scl_slope,anlz-method}
\alias{scl_slope,nifti-method}
\alias{scl_slope-methods}
\alias{scl_slope<-}
\alias{scl_slope<-,anlz-method}
\alias{scl_slope<-,nifti-method}
\title{Extract Image scl_slope attribute}
\usage{
scl_slope(object)

\S4method{scl_slope}{nifti}(object)

\S4method{scl_slope}{anlz}(object)

scl_slope(object) <- value

\S4method{scl_slope}{nifti}(object) <- value

\S4method{scl_slope}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to scl_slope}
}
\description{
Methods that act on the ``scl_slope'' in the NIfTI/ANALYZE header.
}

