% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/session_error.R
\docType{methods}
\name{session_error}
\alias{session_error}
\alias{session_error,anlz-method}
\alias{session_error,nifti-method}
\alias{session_error-methods}
\alias{session_error<-}
\alias{session_error<-,anlz-method}
\alias{session_error<-,nifti-method}
\title{Extract Image session_error attribute}
\usage{
session_error(object)

\S4method{session_error}{nifti}(object)

\S4method{session_error}{anlz}(object)

session_error(object) <- value

\S4method{session_error}{nifti}(object) <- value

\S4method{session_error}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to session_error}
}
\description{
Methods that act on the ``session_error'' in the NIfTI/ANALYZE header.
}

