% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sform_code.R
\docType{methods}
\name{sform_code-methods}
\alias{sform.code}
\alias{sform.code,nifti-method}
\alias{sform.code<-}
\alias{sform.code<-,nifti-method}
\alias{sform_code}
\alias{sform_code,nifti-method}
\alias{sform_code-methods}
\alias{sform_code-methods,}
\alias{sform_code<-}
\alias{sform_code<-,nifti-method}
\title{Extract Image Attribute \code{sform_code}}
\usage{
sform_code(object)

\S4method{sform_code}{nifti}(object)

sform_code(object) <- value

\S4method{sform_code}{nifti}(object) <- value

sform.code(object)

\S4method{sform.code}{nifti}(object)

sform.code(object) <- value

\S4method{sform.code}{nifti}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}.}

\item{value}{is the value to assign to the \code{sform_code} field.}
}
\description{
Methods that act on the \code{sform_code} field in the
NIfTI/ANALYZE header.
}
\details{
See documentation on the ANALYZE and/or NIfTI data standards for
more details.
}
\author{
John Muschelli \email{muschellij2@gmail.com},\cr
Brandon Whitcher \email{bwhitcher@gmail.com}
}
\references{
ANALYZE 7.5\cr
\url{https://rportal.mayo.edu/bir/ANALYZE75.pdf}\cr
NIfTI-1\cr
\url{http://nifti.nimh.nih.gov/}
}

