% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/sizeof_hdr.R
\docType{methods}
\name{sizeof_hdr}
\alias{sizeof_hdr}
\alias{sizeof_hdr,anlz-method}
\alias{sizeof_hdr,nifti-method}
\alias{sizeof_hdr-methods}
\alias{sizeof_hdr<-}
\alias{sizeof_hdr<-,anlz-method}
\alias{sizeof_hdr<-,nifti-method}
\title{Extract Image sizeof_hdr attribute}
\usage{
sizeof_hdr(object)

\S4method{sizeof_hdr}{nifti}(object)

\S4method{sizeof_hdr}{anlz}(object)

sizeof_hdr(object) <- value

\S4method{sizeof_hdr}{nifti}(object) <- value

\S4method{sizeof_hdr}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to sizeof_hdr}
}
\description{
Methods that act on the ``sizeof_hdr'' in the NIfTI/ANALYZE header.
}

