% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/slice_duration.R
\docType{methods}
\name{slice_duration-methods}
\alias{slice_duration}
\alias{slice_duration,anlz-method}
\alias{slice_duration,nifti-method}
\alias{slice_duration-methods}
\alias{slice_duration<-}
\alias{slice_duration<-,anlz-method}
\alias{slice_duration<-,nifti-method}
\title{Extract Image slice_duration attribute}
\usage{
slice_duration(object)

\S4method{slice_duration}{nifti}(object)

\S4method{slice_duration}{anlz}(object)

slice_duration(object) <- value

\S4method{slice_duration}{nifti}(object) <- value

\S4method{slice_duration}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to slice_duration}
}
\description{
Methods that act on the ``slice_duration'' in the NIfTI/ANALYZE header.
}

