% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/slice_end.R
\docType{methods}
\name{slice_end-methods}
\alias{slice_end}
\alias{slice_end,anlz-method}
\alias{slice_end,nifti-method}
\alias{slice_end-methods}
\alias{slice_end<-}
\alias{slice_end<-,anlz-method}
\alias{slice_end<-,nifti-method}
\title{Extract Image slice_end attribute}
\usage{
slice_end(object)

\S4method{slice_end}{nifti}(object)

\S4method{slice_end}{anlz}(object)

slice_end(object) <- value

\S4method{slice_end}{nifti}(object) <- value

\S4method{slice_end}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to slice_end}
}
\description{
Methods that act on the ``slice_end'' in the NIfTI/ANALYZE header.
}

