% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/smax.R
\docType{methods}
\name{smax-methods}
\alias{smax}
\alias{smax,anlz-method}
\alias{smax,nifti-method}
\alias{smax-methods}
\alias{smax<-}
\alias{smax<-,anlz-method}
\alias{smax<-,nifti-method}
\title{Extract Image smax attribute}
\usage{
smax(object)

\S4method{smax}{nifti}(object)

\S4method{smax}{anlz}(object)

smax(object) <- value

\S4method{smax}{nifti}(object) <- value

\S4method{smax}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to smax}
}
\description{
Methods that act on the ``smax'' in the NIfTI/ANALYZE header.
}

