% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/smin.R
\docType{methods}
\name{smin}
\alias{smin}
\alias{smin,anlz-method}
\alias{smin,nifti-method}
\alias{smin-methods}
\alias{smin<-}
\alias{smin<-,anlz-method}
\alias{smin<-,nifti-method}
\title{Extract Image smin attribute}
\usage{
smin(object)

\S4method{smin}{nifti}(object)

\S4method{smin}{anlz}(object)

smin(object) <- value

\S4method{smin}{nifti}(object) <- value

\S4method{smin}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to smin}
}
\description{
Methods that act on the ``smin'' in the NIfTI/ANALYZE header.
}

