% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/srow_x.R
\docType{methods}
\name{srow_x-methods}
\alias{srow_x}
\alias{srow_x,anlz-method}
\alias{srow_x,nifti-method}
\alias{srow_x-methods}
\alias{srow_x<-}
\alias{srow_x<-,anlz-method}
\alias{srow_x<-,nifti-method}
\title{Extract Image srow_x attribute}
\usage{
srow_x(object)

\S4method{srow_x}{nifti}(object)

\S4method{srow_x}{anlz}(object)

srow_x(object) <- value

\S4method{srow_x}{nifti}(object) <- value

\S4method{srow_x}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to srow_x}
}
\description{
Methods that act on the ``srow_x'' in the NIfTI/ANALYZE header.
}

