% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/srow_y.R
\docType{methods}
\name{srow_y-methods}
\alias{srow_y}
\alias{srow_y,anlz-method}
\alias{srow_y,nifti-method}
\alias{srow_y-methods}
\alias{srow_y<-}
\alias{srow_y<-,anlz-method}
\alias{srow_y<-,nifti-method}
\title{Extract Image srow_y attribute}
\usage{
srow_y(object)

\S4method{srow_y}{nifti}(object)

\S4method{srow_y}{anlz}(object)

srow_y(object) <- value

\S4method{srow_y}{nifti}(object) <- value

\S4method{srow_y}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to srow_y}
}
\description{
Methods that act on the ``srow_y'' in the NIfTI/ANALYZE header.
}

