% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/srow_z.R
\docType{methods}
\name{srow_z}
\alias{srow_z}
\alias{srow_z,anlz-method}
\alias{srow_z,nifti-method}
\alias{srow_z-methods}
\alias{srow_z<-}
\alias{srow_z<-,anlz-method}
\alias{srow_z<-,nifti-method}
\title{Extract Image srow_z attribute}
\usage{
srow_z(object)

\S4method{srow_z}{nifti}(object)

\S4method{srow_z}{anlz}(object)

srow_z(object) <- value

\S4method{srow_z}{nifti}(object) <- value

\S4method{srow_z}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to srow_z}
}
\description{
Methods that act on the ``srow_z'' in the NIfTI/ANALYZE header.
}

