% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/start_field.R
\docType{methods}
\name{start_field-methods}
\alias{start_field}
\alias{start_field,anlz-method}
\alias{start_field,nifti-method}
\alias{start_field-methods}
\alias{start_field<-}
\alias{start_field<-,anlz-method}
\alias{start_field<-,nifti-method}
\title{Extract Image start_field attribute}
\usage{
start_field(object)

\S4method{start_field}{nifti}(object)

\S4method{start_field}{anlz}(object)

start_field(object) <- value

\S4method{start_field}{nifti}(object) <- value

\S4method{start_field}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to start_field}
}
\description{
Methods that act on the ``start_field'' in the NIfTI/ANALYZE header.
}

