% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/All_Generics.R
\docType{methods}
\name{[,nifti,missing,missing-method}
\alias{[,nifti,missing,missing,ANY-method}
\alias{[,nifti,missing,missing-method}
\title{Subset a nifti}
\usage{
\S4method{[}{nifti,missing,missing}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{nifti}

\item{i}{boolean or integer index}

\item{j}{boolean or integer index}

\item{...}{boolean or integer index}

\item{drop}{Drop 1-dim elements}
}
\value{
nifti, matrix or array
}
\description{
Returns a subsetted nifti
}

