% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/toffset.R
\docType{methods}
\name{toffset-methods}
\alias{toffset}
\alias{toffset,anlz-method}
\alias{toffset,nifti-method}
\alias{toffset-methods}
\alias{toffset<-}
\alias{toffset<-,anlz-method}
\alias{toffset<-,nifti-method}
\title{Extract Image toffset attribute}
\usage{
toffset(object)

\S4method{toffset}{nifti}(object)

\S4method{toffset}{anlz}(object)

toffset(object) <- value

\S4method{toffset}{nifti}(object) <- value

\S4method{toffset}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to toffset}
}
\description{
Methods that act on the ``toffset'' in the NIfTI/ANALYZE header.
}

