% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/unused1.R
\docType{methods}
\name{unused1-methods}
\alias{unused1}
\alias{unused1,anlz-method}
\alias{unused1,nifti-method}
\alias{unused1-methods}
\alias{unused1<-}
\alias{unused1<-,anlz-method}
\alias{unused1<-,nifti-method}
\title{Extract Image unused1 attribute}
\usage{
unused1(object)

\S4method{unused1}{nifti}(object)

\S4method{unused1}{anlz}(object)

unused1(object) <- value

\S4method{unused1}{nifti}(object) <- value

\S4method{unused1}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to unused1}
}
\description{
Methods that act on the ``unused1'' in the NIfTI/ANALYZE header.
}

