% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/verified.R
\docType{methods}
\name{verified}
\alias{verified}
\alias{verified,anlz-method}
\alias{verified,nifti-method}
\alias{verified-methods}
\alias{verified<-}
\alias{verified<-,anlz-method}
\alias{verified<-,nifti-method}
\title{Extract Image verified attribute}
\usage{
verified(object)

\S4method{verified}{nifti}(object)

\S4method{verified}{anlz}(object)

verified(object) <- value

\S4method{verified}{nifti}(object) <- value

\S4method{verified}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to verified}
}
\description{
Methods that act on the ``verified'' in the NIfTI/ANALYZE header.
}

