% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/views.R
\docType{methods}
\name{views-methods}
\alias{views}
\alias{views,anlz-method}
\alias{views,nifti-method}
\alias{views-methods}
\alias{views<-}
\alias{views<-,anlz-method}
\alias{views<-,nifti-method}
\title{Extract Image views attribute}
\usage{
views(object)

\S4method{views}{nifti}(object)

\S4method{views}{anlz}(object)

views(object) <- value

\S4method{views}{nifti}(object) <- value

\S4method{views}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to views}
}
\description{
Methods that act on the ``views'' in the NIfTI/ANALYZE header.
}

