% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/vols_added.R
\docType{methods}
\name{vols_added-methods}
\alias{vols_added}
\alias{vols_added,anlz-method}
\alias{vols_added,nifti-method}
\alias{vols_added-methods}
\alias{vols_added<-}
\alias{vols_added<-,anlz-method}
\alias{vols_added<-,nifti-method}
\title{Extract Image vols_added attribute}
\usage{
vols_added(object)

\S4method{vols_added}{nifti}(object)

\S4method{vols_added}{anlz}(object)

vols_added(object) <- value

\S4method{vols_added}{nifti}(object) <- value

\S4method{vols_added}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to vols_added}
}
\description{
Methods that act on the ``vols_added'' in the NIfTI/ANALYZE header.
}

