% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vox_offset.R
\docType{methods}
\name{vox_offset-methods}
\alias{vox_offset}
\alias{vox_offset,anlz-method}
\alias{vox_offset,nifti-method}
\alias{vox_offset-methods}
\alias{vox_offset<-}
\alias{vox_offset<-,anlz-method}
\alias{vox_offset<-,nifti-method}
\title{Extract Image vox_offset attribute}
\usage{
vox_offset(object)

\S4method{vox_offset}{nifti}(object)

\S4method{vox_offset}{anlz}(object)

vox_offset(object) <- value

\S4method{vox_offset}{nifti}(object) <- value

\S4method{vox_offset}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to vox_offset}
}
\description{
Methods that act on the ``vox_offset'' in the NIfTI/ANALYZE header.
}

