% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/vox_units.R
\docType{methods}
\name{vox_units-methods}
\alias{vox_units}
\alias{vox_units,anlz-method}
\alias{vox_units,nifti-method}
\alias{vox_units-methods}
\alias{vox_units<-}
\alias{vox_units<-,anlz-method}
\alias{vox_units<-,nifti-method}
\title{Extract Image vox_units attribute}
\usage{
vox_units(object)

\S4method{vox_units}{nifti}(object)

\S4method{vox_units}{anlz}(object)

vox_units(object) <- value

\S4method{vox_units}{nifti}(object) <- value

\S4method{vox_units}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to vox_units}
}
\description{
Methods that act on the ``vox_units'' in the NIfTI/ANALYZE header.
}

