% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voxres.R
\name{voxres}
\alias{voxres}
\title{Gets Voxel Resolution}
\usage{
voxres(img)
}
\arguments{
\item{img}{nifti object}
}
\value{
Scalar numeric, one number, in cubic mm.
}
\description{
Grabs the 3 voxel dimensions and takes the product
}

