% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afniS4.R
\docType{methods}
\name{writeAFNI-methods}
\alias{writeAFNI-methods}
\alias{writeAFNI}
\alias{writeAFNI,afni-method}
\alias{writeAFNI,ANY-method}
\title{writeAFNI}
\usage{
writeAFNI(nim, ...)

\S4method{writeAFNI}{afni}(nim, fname, verbose = FALSE, warn = -1)
}
\arguments{
\item{nim}{is an object of class \code{afni}.}

\item{...}{Additional variables defined by the method.}

\item{fname}{is the path and file name to save the AFNI file (.HEAD/BRIK)
\bold{without} the suffix.}

\item{verbose}{is a logical variable (default = \code{FALSE}) that allows
text-based feedback during execution of the function.}

\item{warn}{is a number to regulate the display of warnings (default = -1).
See \code{\link{options}} for more details.}
}
\value{
Nothing.
}
\description{
This function saves a afni-class object to HEAD/BRIK pair in
AFNI format.
}
\details{
The \code{writeAFNI} function utilizes the internal \code{writeBin}
and \code{writeLines} command to write information to header/binary file
pair.
  
Current acceptable data types include 
\describe{ 
\item{INT16"}{DT SIGNED SHORT (16 bits per voxel)} 
\item{FLOAT32"}{DT FLOAT (32 bits per voxel)} 
\item{"COMPLEX128"}{DT COMPLEX (128 bits per voxel)} 
}
}
\section{Methods}{
 
\describe{ 
\item{nim = "afni"}{Write AFNI volume to disk.} 
\item{nim = "ANY"}{Not implemented.} 
}
}

\examples{

## Taken from the AFNI Matlab Library
## http://afni.nimh.nih.gov/pub/dist/data/afni_matlab_data.tgz
afni.path <- system.file("afni", package="oro.nifti")
orig <- readAFNI(file.path(afni.path, "ARzs_CW_avvr.DEL+orig"))
fname = file.path(tempdir(), "test-afni-image")
writeAFNI(orig, fname, verbose=TRUE)

data <- readAFNI(fname, verbose=TRUE)
image(orig, zlim=c(0.5,256), oma=rep(2,4), bg="white")
image(data, zlim=c(0.5,256), oma=rep(2,4), bg="white")
abs.err <- abs(data - orig)
image(as(abs.err, "nifti"), zlim=range(0,1), oma=rep(2,4),
      bg="white")
}
\references{
AFNI\cr
\url{http://afni.nimh.nih.gov/pub/dist/src/README.attributes}
}
\seealso{
\code{\link{writeANALYZE}}, \code{\link{writeNIfTI}}
}
\author{
Karsten Tabelow \email{karsten.tabelow@wias-berlin.de}
}
\keyword{file}
\keyword{methods}
