% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeS4.R
\name{anlz}
\alias{anlz}
\title{Constructor for Analyze}
\usage{
anlz(img = array(0, dim = rep(1, 4)), dim, datatype = 2, ...)
}
\arguments{
\item{img}{is a multidimensional array of data.}

\item{dim}{is the dimension of the data (default = \code{missing}).}

\item{datatype}{is an integer that denotes the type of data contained in
each voxel.  See the function \code{convert.datatype.anlz} or the 
ANALYZE documentation for more details.}

\item{\dots}{allows for additional \sQuote{slots} to be specified.}
}
\value{
An object of class \code{anlz}.
}
\description{
Constructor for Analyze class objects.
}
\examples{

aim <- anlz() # default

}
\references{
ANALYZE 7.5\cr \url{http://eeg.sourceforge.net/ANALYZE75.pdf}
}
\seealso{
\code{\linkS4class{anlz}}, \code{\linkS4class{nifti}},
\code{\link{nifti}}, \code{\link{convert.datatype.anlz}}
}
\author{
Brandon Whitcher \email{bwhitcher@gmail.com}
}
