% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niftiS4.R
\docType{methods}
\name{audit.trail-methods}
\alias{audit.trail-methods}
\alias{audit.trail,nifti-method}
\alias{audit.trail}
\alias{audit.trail<-,nifti-method}
\alias{audit.trail<-}
\title{Extract or Replace NIfTI Audit Trail}
\usage{
audit.trail(object)

\S4method{audit.trail}{nifti}(object)

audit.trail(object) <- value

\S4method{audit.trail}{nifti}(object) <- value
}
\arguments{
\item{object}{is of class \code{nifti}.}

\item{value}{Value to assign to trail slot}
}
\description{
Operators that act on the audit trail (XML) in the NIfTI header.
}
\section{Methods}{

\describe{
\item{object = "nifti"}{Extract or replace NIfTI audit trail.}
}
}

\author{
Andrew Thornton \email{zeripath@users.sourceforge.net}
}
\keyword{methods}
