% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aux_file.R
\docType{methods}
\name{aux.file-methods}
\alias{aux.file}
\alias{aux.file,anlz-method}
\alias{aux.file,nifti-method}
\alias{aux.file-methods}
\alias{aux.file-methods,}
\alias{aux.file<-}
\alias{aux.file<-,anlz-method}
\alias{aux.file<-,nifti-method}
\title{Extract Image Attribute \code{aux.file}}
\usage{
aux.file(object)

\S4method{aux.file}{nifti}(object)

\S4method{aux.file}{anlz}(object)

aux.file(object) <- value

\S4method{aux.file}{nifti}(object) <- value

\S4method{aux.file}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}.}

\item{value}{is the value to assign to the \code{aux_file} field.}
}
\description{
Methods that act on the \code{aux_file} field in the
NIfTI/ANALYZE header.
}
\examples{
\dontrun{
url <- "http://nifti.nimh.nih.gov/nifti-1/data/avg152T1_RL_nifti.nii.gz"
urlfile <- file.path(system.file("nifti", package="oro.nifti"),
                    "mniRL.nii.gz")
download.file(url, urlfile, quiet=TRUE)
}
options("niftiAuditTrail"=FALSE)

urlfile <- file.path(system.file("nifti", package="oro.nifti"),
                     "mniRL.nii.gz")
mniRL <- readNIfTI(urlfile)
aux.file(mniRL)
aux.file(mniRL) <- "avg152T1_RL_nifti"
aux.file(mniRL)
}
\author{
Brandon Whitcher \email{bwhitcher@gmail.com}
}
\references{
ANALYZE 7.5\cr
\url{http://www.mayo.edu/bir/PDF/ANALYZE75.pdf}\cr
NIfTI-1\cr
\url{http://nifti.nimh.nih.gov/}
}

