% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bitpix.R
\docType{methods}
\name{bitpix-methods}
\alias{bitpix}
\alias{bitpix,anlz-method}
\alias{bitpix,nifti-method}
\alias{bitpix-methods}
\alias{bitpix<-}
\alias{bitpix<-,anlz-method}
\alias{bitpix<-,nifti-method}
\title{Extract Image bitpix attribute}
\usage{
bitpix(object)

\S4method{bitpix}{nifti}(object)

\S4method{bitpix}{anlz}(object)

bitpix(object) <- value

\S4method{bitpix}{nifti}(object) <- value

\S4method{bitpix}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to bitpix}
}
\description{
Methods that act on the ``bitpix'' in the NIfTI/ANALYZE header.
}

