% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bitpix.R
\docType{methods}
\name{bitpix-methods}
\alias{bitpix}
\alias{bitpix,anlz-method}
\alias{bitpix,nifti-method}
\alias{bitpix-methods}
\alias{bitpix-methods,}
\alias{bitpix<-}
\alias{bitpix<-,anlz-method}
\alias{bitpix<-,nifti-method}
\title{Extract Image Attribute \code{bitpix}}
\usage{
bitpix(object)

\S4method{bitpix}{nifti}(object)

\S4method{bitpix}{anlz}(object)

bitpix(object) <- value

\S4method{bitpix}{nifti}(object) <- value

\S4method{bitpix}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}.}

\item{value}{is the value to assign to the \code{bitpix} field.}
}
\description{
Methods that act on the \code{bitpix} field in the
NIfTI/ANALYZE header.
}
\references{
ANALYZE 7.5\cr
\url{http://www.mayo.edu/bir/PDF/ANALYZE75.pdf}\cr
NIfTI-1\cr
\url{http://nifti.nimh.nih.gov/}
}

