% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_units.R
\docType{methods}
\name{cal_units-methods}
\alias{cal_units-methods}
\alias{cal_units-methods,}
\alias{cal_units}
\alias{cal_units,anlz-method}
\alias{cal_units<-}
\alias{cal_units<-,anlz-method}
\alias{cal.units}
\alias{cal.units,nifti-method}
\alias{cal.units,anlz-method}
\alias{cal.units<-}
\alias{cal.units<-,anlz-method}
\title{Extract Image Attribute \code{cal_units}}
\usage{
cal_units(object)

\S4method{cal_units}{anlz}(object)

cal_units(object) <- value

\S4method{cal_units}{anlz}(object) <- value

cal.units(object)

\S4method{cal.units}{anlz}(object)

cal.units(object) <- value

\S4method{cal.units}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}.}

\item{value}{is the value to assign to the \code{cal_units} field.}
}
\description{
Methods that act on the \code{cal_units} field in the
NIfTI/ANALYZE header.
}
\details{
See documentation on the ANALYZE and/or NIfTI data standards for
more details.
}
\references{
ANALYZE 7.5\cr
\url{http://eeg.sourceforge.net/ANALYZE75.pdf}\cr
NIfTI-1\cr
\url{http://nifti.nimh.nih.gov/}
}
\author{
John Muschelli \email{muschellij2@gmail.com},\cr
Brandon Whitcher \email{bwhitcher@gmail.com}
}
