% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_type.R, R/zzz_niftiImage.R
\docType{methods}
\name{data_type-methods}
\alias{data_type-methods}
\alias{data_type-methods,}
\alias{data_type}
\alias{data_type,nifti-method}
\alias{data_type,anlz-method}
\alias{data_type<-}
\alias{data_type<-,nifti-method}
\alias{data_type<-,anlz-method}
\alias{data.type}
\alias{data.type,nifti-method}
\alias{data.type,anlz-method}
\alias{data.type<-}
\alias{data.type<-,nifti-method}
\alias{data.type<-,anlz-method}
\alias{data_type,niftiImage-method}
\alias{datatype,niftiImage-method}
\title{Extract Image Attribute \code{data_type}}
\usage{
data_type(object)

\S4method{data_type}{nifti}(object)

\S4method{data_type}{anlz}(object)

data_type(object) <- value

\S4method{data_type}{nifti}(object) <- value

\S4method{data_type}{anlz}(object) <- value

data.type(object)

\S4method{data.type}{nifti}(object)

\S4method{data.type}{anlz}(object)

data.type(object) <- value

\S4method{data.type}{nifti}(object) <- value

\S4method{data.type}{anlz}(object) <- value

\S4method{data_type}{niftiImage}(object)

\S4method{datatype}{niftiImage}(object)
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}.}

\item{value}{is the value to assign to the \code{data_type} field.}
}
\description{
Methods that act on the \code{data_type} field in the
NIfTI/ANALYZE header.
}
\details{
See documentation on the ANALYZE and/or NIfTI data standards for
more details.
}
\examples{
file = system.file("extdata", "example.nii.gz", package = "RNifti")
img = RNifti::readNifti(file)
data_type(img)
datatype(img)
}
\references{
ANALYZE 7.5\cr
\url{http://eeg.sourceforge.net/ANALYZE75.pdf}\cr
NIfTI-1\cr
\url{http://nifti.nimh.nih.gov/}
}
\author{
John Muschelli \email{muschellij2@gmail.com},\cr
Brandon Whitcher \email{bwhitcher@gmail.com}
}
