% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\docType{methods}
\name{datatype-methods}
\alias{datatype}
\alias{datatype,anlz-method}
\alias{datatype,nifti-method}
\alias{datatype-methods}
\alias{datatype-methods,}
\alias{datatype<-}
\alias{datatype<-,anlz-method}
\alias{datatype<-,nifti-method}
\title{Extract Image Attribute \code{datatype}}
\usage{
datatype(object)

\S4method{datatype}{nifti}(object)

\S4method{datatype}{anlz}(object)

datatype(object) <- value

\S4method{datatype}{nifti}(object) <- value

\S4method{datatype}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}.}

\item{value}{is the value to assign to the \code{datatype} field.}
}
\description{
Methods that act on the \code{datatype} field in the
NIfTI/ANALYZE header.
}
\details{
See documentation on the ANALYZE and/or NIfTI data standards for
more details.
}
\author{
John Muschelli \email{muschellij2@gmail.com},\cr
Brandon Whitcher \email{bwhitcher@gmail.com}
}
\references{
ANALYZE 7.5\cr
\url{https://rportal.mayo.edu/bir/ANALYZE75.pdf}\cr
NIfTI-1\cr
\url{http://nifti.nimh.nih.gov/}
}

