% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_name.R
\docType{methods}
\name{db_name-methods}
\alias{db_name-methods}
\alias{db_name-methods,}
\alias{db_name}
\alias{db_name,nifti-method}
\alias{db_name,anlz-method}
\alias{db_name<-}
\alias{db_name<-,nifti-method}
\alias{db_name<-,anlz-method}
\alias{db.name}
\alias{db.name,nifti-method}
\alias{db.name,anlz-method}
\alias{db.name<-}
\alias{db.name<-,nifti-method}
\alias{db.name<-,anlz-method}
\title{Extract Image Attribute \code{db_name}}
\usage{
db_name(object)

\S4method{db_name}{nifti}(object)

\S4method{db_name}{anlz}(object)

db_name(object) <- value

\S4method{db_name}{nifti}(object) <- value

\S4method{db_name}{anlz}(object) <- value

db.name(object)

\S4method{db.name}{nifti}(object)

\S4method{db.name}{anlz}(object)

db.name(object) <- value

\S4method{db.name}{nifti}(object) <- value

\S4method{db.name}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}.}

\item{value}{is the value to assign to the \code{db_name} field.}
}
\description{
Methods that act on the \code{db_name} field in the
NIfTI/ANALYZE header.
}
\details{
See documentation on the ANALYZE and/or NIfTI data standards for
more details.
}
\references{
ANALYZE 7.5\cr
\url{http://eeg.sourceforge.net/ANALYZE75.pdf}\cr
NIfTI-1\cr
\url{http://nifti.nimh.nih.gov/}
}
\author{
John Muschelli \email{muschellij2@gmail.com},\cr
Brandon Whitcher \email{bwhitcher@gmail.com}
}
