% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/descrip.R
\docType{methods}
\name{descrip-methods}
\alias{descrip}
\alias{descrip,anlz-method}
\alias{descrip,nifti-method}
\alias{descrip-methods}
\alias{descrip<-}
\alias{descrip<-,anlz-method}
\alias{descrip<-,nifti-method}
\title{Extract Image descrip attribute}
\usage{
descrip(object)

\S4method{descrip}{nifti}(object)

\S4method{descrip}{anlz}(object)

descrip(object) <- value

\S4method{descrip}{nifti}(object) <- value

\S4method{descrip}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to descrip}
}
\description{
Methods that act on the ``descrip'' in the NIfTI/ANALYZE header.
}
\examples{
\dontrun{
url <- "http://nifti.nimh.nih.gov/nifti-1/data/avg152T1_LR_nifti.nii.gz"
urlfile <- file.path(system.file("nifti", package="oro.nifti"),
                     "mniLR.nii.gz")
download.file(url, urlfile, quiet=TRUE)
}
urlfile <- file.path(system.file("nifti", package="oro.nifti"),
                     "mniLR.nii.gz")
mniLR <- readNIfTI(urlfile)
descrip(mniLR)
\dontrun{
descrip(mniLR) <- paste(descrip(mniLR), version$version.string, sep="; ")
descrip(mniLR)
}
}

