% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_.R
\docType{methods}
\name{dim_-methods}
\alias{dim_-methods}
\alias{dim_-methods,}
\alias{dim_}
\alias{dim_,nifti-method}
\alias{dim_,anlz-method}
\alias{dim_<-}
\alias{dim_<-,nifti-method}
\alias{dim_<-,anlz-method}
\alias{dim_,ANY-method}
\title{Extract Image Attribute \code{dim_}}
\usage{
dim_(object)

\S4method{dim_}{nifti}(object)

\S4method{dim_}{anlz}(object)

dim_(object) <- value

\S4method{dim_}{nifti}(object) <- value

\S4method{dim_}{anlz}(object) <- value

\S4method{dim_}{ANY}(object)
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}.}

\item{value}{is the value to assign to the \code{dim_} field.}
}
\description{
Methods that act on the \code{dim_} field in the
NIfTI/ANALYZE header.
}
\details{
See documentation on the ANALYZE and/or NIfTI data standards for
more details.
}
\references{
ANALYZE 7.5\cr
\url{http://eeg.sourceforge.net/ANALYZE75.pdf}\cr
NIfTI-1\cr
\url{http://nifti.nimh.nih.gov/}
}
\author{
John Muschelli \email{muschellij2@gmail.com},\cr
Brandon Whitcher \email{bwhitcher@gmail.com}
}
