% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_info.R
\docType{methods}
\name{dim_info-methods}
\alias{dim_info-methods}
\alias{dim_info-methods,}
\alias{dim_info}
\alias{dim_info,nifti-method}
\alias{dim_info<-}
\alias{dim_info<-,nifti-method}
\title{Extract Image Attribute \code{dim_info}}
\usage{
dim_info(object)

\S4method{dim_info}{nifti}(object)

dim_info(object) <- value

\S4method{dim_info}{nifti}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}.}

\item{value}{is the value to assign to the \code{dim_info} field.}
}
\description{
Methods that act on the \code{dim_info} field in the
NIfTI/ANALYZE header.
}
\details{
See documentation on the ANALYZE and/or NIfTI data standards for
more details.
}
\references{
ANALYZE 7.5\cr
\url{http://eeg.sourceforge.net/ANALYZE75.pdf}\cr
NIfTI-1\cr
\url{http://nifti.nimh.nih.gov/}
}
\author{
John Muschelli \email{muschellij2@gmail.com},\cr
Brandon Whitcher \email{bwhitcher@gmail.com}
}
