% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/drop_img_dim.R
\name{drop_img_dim}
\alias{drop_img_dim}
\title{Drop Image Dimension}
\usage{
drop_img_dim(img, warn = TRUE)
}
\arguments{
\item{img}{nifti object}

\item{warn}{output warning if dimensions are under 3}
}
\value{
Object of class nifti
}
\description{
Drops a dimension of an image that has one-dimension and
sets respective values to 0 in pixdim or 1 in dim
}

