% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/drop_img_dim.R
\name{drop_img_dim}
\alias{drop_img_dim}
\title{Drop Image Dimension}
\usage{
drop_img_dim(img, onlylast = TRUE, warn = TRUE)
}
\arguments{
\item{img}{nifti object}

\item{onlylast}{Drop dimension only if it's the last dimensions.
For example if dim is 10x10x1x10 then no dimension is dropped, but
if 10x10x10x1 then it will be 10x10x10}

\item{warn}{output warning if dimensions are under 3}
}
\value{
Object of class nifti
}
\description{
Drops a dimension of an image that has one-dimension and
sets respective values to 0 in pixdim or 1 in dim
}
\examples{
nim = nifti(array(rnorm(10^3), dim= rep(10, 3)))
nim2 = nifti(array(rnorm(10^3), dim= c(10, 10, 1, 10)))
drop_img_dim(nim2)
drop_img_dim(nim2, onlylast = FALSE)
nim3 = nifti(array(rnorm(10^3), dim= c(10, 10, 10, 1)))
drop_img_dim(nim3)
drop_img_dim(nim3, onlylast = FALSE) # the same as above
nim4 = nifti(array(rnorm(10^3), dim= c(10, 10, 10, 1, 10)))
dim(nim4[,,,1,])
dim(nim4[,,,1,,drop=TRUE])
drop_img_dim(nim4)

nim5 = nifti(array(rnorm(10^4), dim= c(1, 10, 10, 10, 1, 10)))
drop_img_dim(nim5)
drop_img_dim(nim5, onlylast = FALSE)

nim6 = nifti(array(rnorm(10^3), dim= c(1, 10, 10, 10, 1, 1)))
drop_img_dim(nim6)
\dontrun{
## 27 scans of Colin Holmes (MNI) brain co-registered and averaged
## NIfTI two-file format
URL <- "http://imaging.mrc-cbu.cam.ac.uk/downloads/Colin/colin_1mm.tgz"
urlfile <- file.path(tempdir(), "colin_1mm.tgz")
download.file(URL, dest=urlfile, quiet=TRUE)
untar(urlfile, exdir=tempdir())
colin <- readNIfTI(file.path(tempdir(), "colin_1mm"))
dim(colin)
dim_(colin)
pixdim(colin)
# this will error
writeNIfTI(colin, filename = tempfile())
colin = drop_img_dim(colin)
writeNIfTI(colin, filename = tempfile())
}
}

