% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_time.R
\docType{methods}
\name{exp_time-methods}
\alias{exp_time-methods}
\alias{exp_time-methods,}
\alias{exp_time}
\alias{exp_time,anlz-method}
\alias{exp_time<-}
\alias{exp_time<-,anlz-method}
\title{Extract Image Attribute \code{exp_time}}
\usage{
exp_time(object)

\S4method{exp_time}{anlz}(object)

exp_time(object) <- value

\S4method{exp_time}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}.}

\item{value}{is the value to assign to the \code{exp_time} field.}
}
\description{
Methods that act on the \code{exp_time} field in the
NIfTI/ANALYZE header.
}
\details{
See documentation on the ANALYZE and/or NIfTI data standards for
more details.
}
\references{
ANALYZE 7.5\cr
\url{http://eeg.sourceforge.net/ANALYZE75.pdf}\cr
NIfTI-1\cr
\url{http://nifti.nimh.nih.gov/}
}
\author{
John Muschelli \email{muschellij2@gmail.com},\cr
Brandon Whitcher \email{bwhitcher@gmail.com}
}
