% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extender.R
\docType{methods}
\name{extender-methods}
\alias{extender-methods}
\alias{extender-methods,}
\alias{extender}
\alias{extender,nifti-method}
\alias{extender<-}
\alias{extender<-,nifti-method}
\title{Extract Image Attribute \code{extender}}
\usage{
extender(object)

\S4method{extender}{nifti}(object)

extender(object) <- value

\S4method{extender}{nifti}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}.}

\item{value}{is the value to assign to the \code{extender} field.}
}
\description{
Methods that act on the \code{extender} field in the
NIfTI/ANALYZE header.
}
\details{
See documentation on the ANALYZE and/or NIfTI data standards for
more details.
}
\references{
ANALYZE 7.5\cr
\url{http://eeg.sourceforge.net/ANALYZE75.pdf}\cr
NIfTI-1\cr
\url{http://nifti.nimh.nih.gov/}
}
\author{
John Muschelli \email{muschellij2@gmail.com},\cr
Brandon Whitcher \email{bwhitcher@gmail.com}
}
