% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extents.R
\docType{methods}
\name{extents-methods}
\alias{extents-methods}
\alias{extents-methods,}
\alias{extents}
\alias{extents,nifti-method}
\alias{extents,anlz-method}
\alias{extents<-}
\alias{extents<-,nifti-method}
\alias{extents<-,anlz-method}
\title{Extract Image Attribute \code{extents}}
\usage{
extents(object)

\S4method{extents}{nifti}(object)

\S4method{extents}{anlz}(object)

extents(object) <- value

\S4method{extents}{nifti}(object) <- value

\S4method{extents}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}.}

\item{value}{is the value to assign to the \code{extents} field.}
}
\description{
Methods that act on the \code{extents} field in the
NIfTI/ANALYZE header.
}
\details{
See documentation on the ANALYZE and/or NIfTI data standards for
more details.
}
\references{
ANALYZE 7.5\cr
\url{http://eeg.sourceforge.net/ANALYZE75.pdf}\cr
NIfTI-1\cr
\url{http://nifti.nimh.nih.gov/}
}
\author{
John Muschelli \email{muschellij2@gmail.com},\cr
Brandon Whitcher \email{bwhitcher@gmail.com}
}
