% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_skip.R
\docType{methods}
\name{field_skip-methods}
\alias{field.skip}
\alias{field.skip,anlz-method}
\alias{field.skip,nifti-method}
\alias{field.skip<-}
\alias{field.skip<-,anlz-method}
\alias{field_skip}
\alias{field_skip,anlz-method}
\alias{field_skip-methods}
\alias{field_skip-methods,}
\alias{field_skip<-}
\alias{field_skip<-,anlz-method}
\title{Extract Image Attribute \code{field_skip}}
\usage{
field_skip(object)

\S4method{field_skip}{anlz}(object)

field_skip(object) <- value

\S4method{field_skip}{anlz}(object) <- value

field.skip(object)

\S4method{field.skip}{anlz}(object)

field.skip(object) <- value

\S4method{field.skip}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}.}

\item{value}{is the value to assign to the \code{field_skip} field.}
}
\description{
Methods that act on the \code{field_skip} field in the
NIfTI/ANALYZE header.
}
\details{
See documentation on the ANALYZE and/or NIfTI data standards for
more details.
}
\author{
John Muschelli \email{muschellij2@gmail.com},\cr
Brandon Whitcher \email{bwhitcher@gmail.com}
}
\references{
ANALYZE 7.5\cr
\url{https://rportal.mayo.edu/bir/ANALYZE75.pdf}\cr
NIfTI-1\cr
\url{http://nifti.nimh.nih.gov/}
}

