% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funused2.R
\docType{methods}
\name{funused2-methods}
\alias{funused2-methods}
\alias{funused2-methods,}
\alias{funused2}
\alias{funused2,anlz-method}
\alias{funused2<-}
\alias{funused2<-,anlz-method}
\title{Extract Image Attribute \code{funused2}}
\usage{
funused2(object)

\S4method{funused2}{anlz}(object)

funused2(object) <- value

\S4method{funused2}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}.}

\item{value}{is the value to assign to the \code{funused2} field.}
}
\description{
Methods that act on the \code{funused2} field in the
NIfTI/ANALYZE header.
}
\details{
See documentation on the ANALYZE and/or NIfTI data standards for
more details.
}
\references{
ANALYZE 7.5\cr
\url{http://eeg.sourceforge.net/ANALYZE75.pdf}\cr
NIfTI-1\cr
\url{http://nifti.nimh.nih.gov/}
}
\author{
John Muschelli \email{muschellij2@gmail.com},\cr
Brandon Whitcher \email{bwhitcher@gmail.com}
}
