% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funused3.R
\docType{methods}
\name{funused3-methods}
\alias{funused3-methods}
\alias{funused3-methods,}
\alias{funused3}
\alias{funused3,anlz-method}
\alias{funused3<-}
\alias{funused3<-,anlz-method}
\title{Extract Image Attribute \code{funused3}}
\usage{
funused3(object)

\S4method{funused3}{anlz}(object)

funused3(object) <- value

\S4method{funused3}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}.}

\item{value}{is the value to assign to the \code{funused3} field.}
}
\description{
Methods that act on the \code{funused3} field in the
NIfTI/ANALYZE header.
}
\details{
See documentation on the ANALYZE and/or NIfTI data standards for
more details.
}
\references{
ANALYZE 7.5\cr
\url{http://eeg.sourceforge.net/ANALYZE75.pdf}\cr
NIfTI-1\cr
\url{http://nifti.nimh.nih.gov/}
}
\author{
John Muschelli \email{muschellij2@gmail.com},\cr
Brandon Whitcher \email{bwhitcher@gmail.com}
}
