% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_un0.R
\docType{methods}
\name{hist_un0-methods}
\alias{hist_un0-methods}
\alias{hist_un0-methods,}
\alias{hist_un0}
\alias{hist_un0,anlz-method}
\alias{hist_un0<-}
\alias{hist_un0<-,anlz-method}
\title{Extract Image Attribute \code{hist_un0}}
\usage{
hist_un0(object)

\S4method{hist_un0}{anlz}(object)

hist_un0(object) <- value

\S4method{hist_un0}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}.}

\item{value}{is the value to assign to the \code{hist_un0} field.}
}
\description{
Methods that act on the \code{hist_un0} field in the
NIfTI/ANALYZE header.
}
\details{
See documentation on the ANALYZE and/or NIfTI data standards for
more details.
}
\references{
ANALYZE 7.5\cr
\url{http://eeg.sourceforge.net/ANALYZE75.pdf}\cr
NIfTI-1\cr
\url{http://nifti.nimh.nih.gov/}
}
\author{
John Muschelli \email{muschellij2@gmail.com},\cr
Brandon Whitcher \email{bwhitcher@gmail.com}
}
