% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{integerTranslation}
\alias{integerTranslation}
\alias{invertIntegerTranslation}
\title{integerTranslation}
\usage{
integerTranslation(nim, data, verbose = FALSE)

invertIntegerTranslation(nim, verbose = FALSE)
}
\arguments{
\item{nim}{is an object of class \code{nifti}.}

\item{data}{is ...}

\item{verbose}{is a logical variable (default = \code{FALSE}) that allows
text-based feedback during execution of the function.}
}
\value{
...
}
\description{
...
}
\details{
...
}
\author{
Andrew Thornton \email{zeripath@users.sourceforge.net}
}
