% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/intent_code.R
\docType{methods}
\name{intent_code}
\alias{intent_code}
\alias{intent_code,anlz-method}
\alias{intent_code,nifti-method}
\alias{intent_code-methods}
\alias{intent_code<-}
\alias{intent_code<-,anlz-method}
\alias{intent_code<-,nifti-method}
\title{Extract Image intent_code attribute}
\usage{
intent_code(object)

\S4method{intent_code}{nifti}(object)

\S4method{intent_code}{anlz}(object)

intent_code(object) <- value

\S4method{intent_code}{nifti}(object) <- value

\S4method{intent_code}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to intent_code}
}
\description{
Methods that act on the ``intent_code'' in the NIfTI/ANALYZE header.
}

