% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intent_code.R
\docType{methods}
\name{intent_code-methods}
\alias{intent_code-methods}
\alias{intent_code-methods,}
\alias{intent_code}
\alias{intent_code,nifti-method}
\alias{intent_code<-}
\alias{intent_code<-,nifti-method}
\alias{intent.code}
\alias{intent.code,nifti-method}
\alias{intent.code<-}
\alias{intent.code<-,nifti-method}
\title{Extract Image Attribute \code{intent_code}}
\usage{
intent_code(object)

\S4method{intent_code}{nifti}(object)

intent_code(object) <- value

\S4method{intent_code}{nifti}(object) <- value

intent.code(object)

\S4method{intent.code}{nifti}(object)

intent.code(object) <- value

\S4method{intent.code}{nifti}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}.}

\item{value}{is the value to assign to the \code{intent_code} field.}
}
\description{
Methods that act on the \code{intent_code} field in the
NIfTI/ANALYZE header.
}
\details{
See documentation on the ANALYZE and/or NIfTI data standards for
more details.
}
\references{
ANALYZE 7.5\cr
\url{https://rportal.mayo.edu/bir/ANALYZE75.pdf}\cr
NIfTI-1\cr
\url{http://nifti.nimh.nih.gov/}
}
\author{
John Muschelli \email{muschellij2@gmail.com},\cr
Brandon Whitcher \email{bwhitcher@gmail.com}
}
