% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intent_name.R, R/zzz_niftiImage.R
\docType{methods}
\name{intent_name-methods}
\alias{intent_name-methods}
\alias{intent_name-methods,}
\alias{intent_name}
\alias{intent_name,nifti-method}
\alias{intent_name<-}
\alias{intent_name<-,nifti-method}
\alias{intent.name}
\alias{intent.name,nifti-method}
\alias{intent.name<-}
\alias{intent.name<-,nifti-method}
\alias{intent_name,niftiImage-method}
\alias{intent_name<-,niftiImage-method}
\title{Extract Image Attribute \code{intent_name}}
\usage{
intent_name(object)

\S4method{intent_name}{nifti}(object)

intent_name(object) <- value

\S4method{intent_name}{nifti}(object) <- value

intent.name(object)

\S4method{intent.name}{nifti}(object)

intent.name(object) <- value

\S4method{intent.name}{nifti}(object) <- value

\S4method{intent_name}{niftiImage}(object)

\S4method{intent_name}{niftiImage}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}.}

\item{value}{is the value to assign to the \code{intent_name} field.}
}
\description{
Methods that act on the \code{intent_name} field in the
NIfTI/ANALYZE header.
}
\details{
See documentation on the ANALYZE and/or NIfTI data standards for
more details.
}
\examples{
file = system.file("extdata", "example.nii.gz", package = "RNifti")
img = RNifti::readNifti(file)
intent_name(img)
intent_name(img) = "hey"
stopifnot(intent_name(img) ==  "hey")
}
\references{
ANALYZE 7.5\cr
\url{http://eeg.sourceforge.net/ANALYZE75.pdf}\cr
NIfTI-1\cr
\url{http://nifti.nimh.nih.gov/}
}
\author{
John Muschelli \email{muschellij2@gmail.com},\cr
Brandon Whitcher \email{bwhitcher@gmail.com}
}
