% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intent_p1.R, R/zzz_niftiImage.R
\docType{methods}
\name{intent_p1-methods}
\alias{intent_p1-methods}
\alias{intent_p1-methods,}
\alias{intent_p1}
\alias{intent_p1,nifti-method}
\alias{intent_p1<-}
\alias{intent_p1<-,nifti-method}
\alias{intent.p1}
\alias{intent.p1,nifti-method}
\alias{intent.p1<-}
\alias{intent.p1<-,nifti-method}
\alias{intent_p1,niftiImage-method}
\alias{intent_p1<-,niftiImage-method}
\title{Extract Image Attribute \code{intent_p1}}
\usage{
intent_p1(object)

\S4method{intent_p1}{nifti}(object)

intent_p1(object) <- value

\S4method{intent_p1}{nifti}(object) <- value

intent.p1(object)

\S4method{intent.p1}{nifti}(object)

intent.p1(object) <- value

\S4method{intent.p1}{nifti}(object) <- value

\S4method{intent_p1}{niftiImage}(object)

\S4method{intent_p1}{niftiImage}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}.}

\item{value}{is the value to assign to the \code{intent_p1} field.}
}
\description{
Methods that act on the \code{intent_p1} field in the
NIfTI/ANALYZE header.
}
\details{
See documentation on the ANALYZE and/or NIfTI data standards for
more details.
}
\examples{
file = system.file("extdata", "example.nii.gz", package = "RNifti")
img = RNifti::readNifti(file)
intent_p1(img)
intent_p1(img) = 2
stopifnot(intent_p1(img) == 2)
intent_p2(img)
intent_p2(img) = 2
stopifnot(intent_p2(img) == 2)
intent_p3(img)
intent_p3(img) = 2
stopifnot(intent_p3(img) == 2)
}
\references{
ANALYZE 7.5\cr
\url{http://eeg.sourceforge.net/ANALYZE75.pdf}\cr
NIfTI-1\cr
\url{http://nifti.nimh.nih.gov/}
}
\author{
John Muschelli \email{muschellij2@gmail.com},\cr
Brandon Whitcher \email{bwhitcher@gmail.com}
}
