% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intent_p2.R, R/zzz_niftiImage.R
\docType{methods}
\name{intent_p2-methods}
\alias{intent_p2-methods}
\alias{intent_p2-methods,}
\alias{intent_p2}
\alias{intent_p2,nifti-method}
\alias{intent_p2<-}
\alias{intent_p2<-,nifti-method}
\alias{intent.p2}
\alias{intent.p2,nifti-method}
\alias{intent.p2<-}
\alias{intent.p2<-,nifti-method}
\alias{intent_p2,niftiImage-method}
\alias{intent_p2<-,niftiImage-method}
\title{Extract Image Attribute \code{intent_p2}}
\usage{
intent_p2(object)

\S4method{intent_p2}{nifti}(object)

intent_p2(object) <- value

\S4method{intent_p2}{nifti}(object) <- value

intent.p2(object)

\S4method{intent.p2}{nifti}(object)

intent.p2(object) <- value

\S4method{intent.p2}{nifti}(object) <- value

\S4method{intent_p2}{niftiImage}(object)

\S4method{intent_p2}{niftiImage}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}.}

\item{value}{is the value to assign to the \code{intent_p2} field.}
}
\description{
Methods that act on the \code{intent_p2} field in the
NIfTI/ANALYZE header.
}
\details{
See documentation on the ANALYZE and/or NIfTI data standards for
more details.
}
\references{
ANALYZE 7.5\cr
\url{http://eeg.sourceforge.net/ANALYZE75.pdf}\cr
NIfTI-1\cr
\url{http://nifti.nimh.nih.gov/}
}
\author{
John Muschelli \email{muschellij2@gmail.com},\cr
Brandon Whitcher \email{bwhitcher@gmail.com}
}
