% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic.R
\docType{methods}
\name{magic-methods}
\alias{magic}
\alias{magic,nifti-method}
\alias{magic-methods}
\alias{magic-methods,}
\alias{magic<-}
\alias{magic<-,nifti-method}
\title{Extract Image Attribute \code{magic}}
\usage{
magic(object)

\S4method{magic}{nifti}(object)

magic(object) <- value

\S4method{magic}{nifti}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}.}

\item{value}{is the value to assign to the \code{magic} field.}
}
\description{
Methods that act on the \code{magic} field in the
NIfTI/ANALYZE header.
}
\details{
See documentation on the ANALYZE and/or NIfTI data standards for
more details.
}
\author{
John Muschelli \email{muschellij2@gmail.com},\cr
Brandon Whitcher \email{bwhitcher@gmail.com}
}
\references{
ANALYZE 7.5\cr
\url{https://rportal.mayo.edu/bir/ANALYZE75.pdf}\cr
NIfTI-1\cr
\url{http://nifti.nimh.nih.gov/}
}

