% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niftiS4.R
\docType{class}
\name{niftiAuditTrail-class}
\alias{niftiAuditTrail-class}
\title{Class "niftiAuditTrail"}
\description{
An extension of the NIfTI class that adds an audit trail in XML format.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("niftiAuditTrail", data, dim, dimnames, ...)}.
}

\section{Methods}{

  \describe{
    \item{show}{\code{signature(object = "niftiAuditTrail")}: prints out a
                summary of the imaging data.}
  }
}

\examples{

showClass("niftiAuditTrail")

}
\references{
NIfTI-1\cr \url{http://nifti.nimh.nih.gov/}
}
\seealso{
\code{\linkS4class{nifti}}, \code{\linkS4class{niftiExtension}}
}
\author{
Andrew Thornton \email{zeripath@users.sourceforge.net}
}
\keyword{classes}
