% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niftiS4.R
\docType{class}
\name{niftiExtensionSection-class}
\alias{niftiExtensionSection-class}
\title{Class "niftiExtensionSection"}
\description{
A \code{niftiExtensionSection} contains the fields that conform
  to the NIfTI standard regarding header extensions.  A \code{niftiExtension}
  is composed of one or more of these objects.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("niftiExtensionSection", data, dim, dimnames, ...)}.
}

\examples{

showClass("niftiExtensionSection")
}
\references{
NIfTI-1\cr \url{http://nifti.nimh.nih.gov/}
}
\seealso{
\code{\linkS4class{niftiExtension}}, \code{\linkS4class{nifti}}
}
\author{
Brandon Whitcher \email{bwhitcher@gmail.com},\cr Andrew Thornton
\email{zeripath@users.sourcefore.net}
}
\keyword{classes}
