% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Operators.R
\name{nifti-operators}
\alias{nifti-operators}
\alias{Ops,nifti,nifti-method}
\alias{Ops,nifti,numeric-method}
\alias{Ops,numeric,nifti-method}
\title{Operations for NIfTI Objects}
\usage{
\S4method{Ops}{nifti,nifti}(e1, e2)

\S4method{Ops}{nifti,numeric}(e1, e2)

\S4method{Ops}{numeric,nifti}(e1, e2)
}
\arguments{
\item{e1}{is an object of class \code{nifti}.}

\item{e2}{is an object of class \code{nifti}.}
}
\description{
Overloaded operators for nifti objects
}
\examples{

img01 <- nifti(array(1:64, c(4,4,4,1)), datatype=4)
img02 <- nifti(array(64:1, c(4,4,4,1)), datatype=4)
is.nifti(img01 + img02)
is.nifti(sqrt(2) * img01)
is.nifti(img02 / pi)

}
\author{
John Muschelli \email{muschellij2@gmail.com}
}
