% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omin.R
\docType{methods}
\name{omin-methods}
\alias{omin-methods}
\alias{omin-methods,}
\alias{omin}
\alias{omin,anlz-method}
\alias{omin<-}
\alias{omin<-,anlz-method}
\title{Extract Image Attribute \code{omin}}
\usage{
omin(object)

\S4method{omin}{anlz}(object)

omin(object) <- value

\S4method{omin}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}.}

\item{value}{is the value to assign to the \code{omin} field.}
}
\description{
Methods that act on the \code{omin} field in the
NIfTI/ANALYZE header.
}
\details{
See documentation on the ANALYZE and/or NIfTI data standards for
more details.
}
\references{
ANALYZE 7.5\cr
\url{http://eeg.sourceforge.net/ANALYZE75.pdf}\cr
NIfTI-1\cr
\url{http://nifti.nimh.nih.gov/}
}
\author{
John Muschelli \email{muschellij2@gmail.com},\cr
Brandon Whitcher \email{bwhitcher@gmail.com}
}
