% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/origin.R
\docType{methods}
\name{origin-methods}
\alias{origin}
\alias{origin,anlz-method}
\alias{origin-methods}
\alias{origin-methods,}
\alias{origin<-}
\alias{origin<-,anlz-method}
\title{Extract Image Attribute \code{origin}}
\usage{
origin(object)

\S4method{origin}{anlz}(object)

origin(object) <- value

\S4method{origin}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}.}

\item{value}{is the value to assign to the \code{origin} field.}
}
\description{
Methods that act on the \code{origin} field in the
NIfTI/ANALYZE header.
}
\details{
See documentation on the ANALYZE and/or NIfTI data standards for
more details.
}
\author{
John Muschelli \email{muschellij2@gmail.com},\cr
Brandon Whitcher \email{bwhitcher@gmail.com}
}
\references{
ANALYZE 7.5\cr
\url{https://rportal.mayo.edu/bir/ANALYZE75.pdf}\cr
NIfTI-1\cr
\url{http://nifti.nimh.nih.gov/}
}

