% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oro2nii.R
\name{oro2nii}
\alias{oro2nii}
\title{Convert oro.nifti nifti to RNifti niftiImage object}
\usage{
oro2nii(image, verbose = FALSE)
}
\arguments{
\item{image}{\code{\link{nifti}} object}

\item{verbose}{print messages, passed to \code{\link{writeNIfTI}}}
}
\value{
Object of class \code{niftiImage}
}
\description{
Converts a nifti from oro.nifti to a 
niftiImage object from the RNifti package
}
