% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_id.R
\docType{methods}
\name{patient_id-methods}
\alias{patient_id-methods}
\alias{patient_id-methods,}
\alias{patient_id}
\alias{patient_id,anlz-method}
\alias{patient_id<-}
\alias{patient_id<-,anlz-method}
\alias{patient.id}
\alias{patient.id,nifti-method}
\alias{patient.id,anlz-method}
\alias{patient.id<-}
\alias{patient.id<-,anlz-method}
\title{Extract Image Attribute \code{patient_id}}
\usage{
patient_id(object)

\S4method{patient_id}{anlz}(object)

patient_id(object) <- value

\S4method{patient_id}{anlz}(object) <- value

patient.id(object)

\S4method{patient.id}{anlz}(object)

patient.id(object) <- value

\S4method{patient.id}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}.}

\item{value}{is the value to assign to the \code{patient_id} field.}
}
\description{
Methods that act on the \code{patient_id} field in the
NIfTI/ANALYZE header.
}
\details{
See documentation on the ANALYZE and/or NIfTI data standards for
more details.
}
\references{
ANALYZE 7.5\cr
\url{http://eeg.sourceforge.net/ANALYZE75.pdf}\cr
NIfTI-1\cr
\url{http://nifti.nimh.nih.gov/}
}
\author{
John Muschelli \email{muschellij2@gmail.com},\cr
Brandon Whitcher \email{bwhitcher@gmail.com}
}
