% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/pixdim.R
\docType{methods}
\name{pixdim}
\alias{pixdim}
\alias{pixdim,anlz-method}
\alias{pixdim,nifti-method}
\alias{pixdim-methods}
\alias{pixdim<-}
\alias{pixdim<-,anlz-method}
\alias{pixdim<-,nifti-method}
\title{Extract Image pixdim attribute}
\usage{
pixdim(object)

\S4method{pixdim}{nifti}(object)

\S4method{pixdim}{anlz}(object)

pixdim(object) <- value

\S4method{pixdim}{nifti}(object) <- value

\S4method{pixdim}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to pixdim}
}
\description{
Methods that act on the ``pixdim'' in the NIfTI/ANALYZE header.
}
\examples{
\dontrun{
url <- "http://nifti.nimh.nih.gov/nifti-1/data/avg152T1_LR_nifti.nii.gz"
urlfile <- file.path(system.file("nifti", package="oro.nifti"),
"mniLR.nii.gz")
download.file(url, urlfile, quiet=TRUE)
}
urlfile <- file.path(system.file("nifti", package="oro.nifti"),
                     "mniLR.nii.gz")
mniLR <- readNIfTI(urlfile)
pixdim(mniLR)
}

