% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quatern_b.R, R/zzz_niftiImage.R
\docType{methods}
\name{quatern_b-methods}
\alias{quatern_b-methods}
\alias{quatern_b-methods,}
\alias{quatern_b}
\alias{quatern_b,nifti-method}
\alias{quatern_b<-}
\alias{quatern_b<-,nifti-method}
\alias{quatern.b}
\alias{quatern.b,nifti-method}
\alias{quatern.b<-}
\alias{quatern.b<-,nifti-method}
\alias{quatern_b,niftiImage-method}
\alias{quatern_b<-,niftiImage-method}
\title{Extract Image Attribute \code{quatern_b}}
\usage{
quatern_b(object)

\S4method{quatern_b}{nifti}(object)

quatern_b(object) <- value

\S4method{quatern_b}{nifti}(object) <- value

quatern.b(object)

\S4method{quatern.b}{nifti}(object)

quatern.b(object) <- value

\S4method{quatern.b}{nifti}(object) <- value

\S4method{quatern_b}{niftiImage}(object)

\S4method{quatern_b}{niftiImage}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}.}

\item{value}{is the value to assign to the \code{quatern_b} field.}
}
\description{
Methods that act on the \code{quatern_b} field in the
NIfTI/ANALYZE header.
}
\details{
See documentation on the ANALYZE and/or NIfTI data standards for
more details.
}
\examples{
file = system.file("extdata", "example.nii.gz", package = "RNifti")
img = RNifti::readNifti(file)
quatern_b(img)
quatern_b(img) = 3
stopifnot(quatern_b(img) == 3)
quatern_c(img)
quatern_c(img) = 3
stopifnot(quatern_c(img) == 3)
quatern_d(img)
quatern_d(img) = 3
stopifnot(quatern_d(img) == 3)
}
\references{
ANALYZE 7.5\cr
\url{http://eeg.sourceforge.net/ANALYZE75.pdf}\cr
NIfTI-1\cr
\url{http://nifti.nimh.nih.gov/}
}
\author{
John Muschelli \email{muschellij2@gmail.com},\cr
Brandon Whitcher \email{bwhitcher@gmail.com}
}
