% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quatern_c.R, R/zzz_niftiImage.R
\docType{methods}
\name{quatern_c-methods}
\alias{quatern_c-methods}
\alias{quatern_c-methods,}
\alias{quatern_c}
\alias{quatern_c,nifti-method}
\alias{quatern_c<-}
\alias{quatern_c<-,nifti-method}
\alias{quatern.c}
\alias{quatern.c,nifti-method}
\alias{quatern.c<-}
\alias{quatern.c<-,nifti-method}
\alias{quatern_c,niftiImage-method}
\alias{quatern_c<-,niftiImage-method}
\title{Extract Image Attribute \code{quatern_c}}
\usage{
quatern_c(object)

\S4method{quatern_c}{nifti}(object)

quatern_c(object) <- value

\S4method{quatern_c}{nifti}(object) <- value

quatern.c(object)

\S4method{quatern.c}{nifti}(object)

quatern.c(object) <- value

\S4method{quatern.c}{nifti}(object) <- value

\S4method{quatern_c}{niftiImage}(object)

\S4method{quatern_c}{niftiImage}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}.}

\item{value}{is the value to assign to the \code{quatern_c} field.}
}
\description{
Methods that act on the \code{quatern_c} field in the
NIfTI/ANALYZE header.
}
\details{
See documentation on the ANALYZE and/or NIfTI data standards for
more details.
}
\references{
ANALYZE 7.5\cr
\url{http://eeg.sourceforge.net/ANALYZE75.pdf}\cr
NIfTI-1\cr
\url{http://nifti.nimh.nih.gov/}
}
\author{
John Muschelli \email{muschellij2@gmail.com},\cr
Brandon Whitcher \email{bwhitcher@gmail.com}
}
