% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quatern_d.R, R/zzz_niftiImage.R
\docType{methods}
\name{quatern_d-methods}
\alias{quatern_d-methods}
\alias{quatern_d-methods,}
\alias{quatern_d}
\alias{quatern_d,nifti-method}
\alias{quatern_d<-}
\alias{quatern_d<-,nifti-method}
\alias{quatern.d}
\alias{quatern.d,nifti-method}
\alias{quatern.d<-}
\alias{quatern.d<-,nifti-method}
\alias{quatern_d,niftiImage-method}
\alias{quatern_d<-,niftiImage-method}
\title{Extract Image Attribute \code{quatern_d}}
\usage{
quatern_d(object)

\S4method{quatern_d}{nifti}(object)

quatern_d(object) <- value

\S4method{quatern_d}{nifti}(object) <- value

quatern.d(object)

\S4method{quatern.d}{nifti}(object)

quatern.d(object) <- value

\S4method{quatern.d}{nifti}(object) <- value

\S4method{quatern_d}{niftiImage}(object)

\S4method{quatern_d}{niftiImage}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}.}

\item{value}{is the value to assign to the \code{quatern_d} field.}
}
\description{
Methods that act on the \code{quatern_d} field in the
NIfTI/ANALYZE header.
}
\details{
See documentation on the ANALYZE and/or NIfTI data standards for
more details.
}
\references{
ANALYZE 7.5\cr
\url{http://eeg.sourceforge.net/ANALYZE75.pdf}\cr
NIfTI-1\cr
\url{http://nifti.nimh.nih.gov/}
}
\author{
John Muschelli \email{muschellij2@gmail.com},\cr
Brandon Whitcher \email{bwhitcher@gmail.com}
}
