% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/regular.R
\docType{methods}
\name{regular-methods}
\alias{regular}
\alias{regular,anlz-method}
\alias{regular,nifti-method}
\alias{regular-methods}
\alias{regular-methods,}
\alias{regular<-}
\alias{regular<-,anlz-method}
\alias{regular<-,nifti-method}
\title{Extract Image Attribute \code{regular}}
\usage{
regular(object)

\S4method{regular}{nifti}(object)

\S4method{regular}{anlz}(object)

regular(object) <- value

\S4method{regular}{nifti}(object) <- value

\S4method{regular}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}.}

\item{value}{is the value to assign to the \code{regular} field.}
}
\description{
Methods that act on the \code{regular} field in the
NIfTI/ANALYZE header.
}
\details{
See documentation on the ANALYZE and/or NIfTI data standards for
more details.
}
\author{
John Muschelli \email{muschellij2@gmail.com},\cr
Brandon Whitcher \email{bwhitcher@gmail.com}
}
\references{
ANALYZE 7.5\cr
\url{https://rportal.mayo.edu/bir/ANALYZE75.pdf}\cr
NIfTI-1\cr
\url{http://nifti.nimh.nih.gov/}
}

