\name{rmniigz}
\alias{rmniigz}
\alias{rmnii}
\alias{rmgz}
\title{Remove File Extensions Around the NIfTI Format}
\description{
  Simple function(s) that remove file extensions commonly found when
  using NIfTI-1 format files.
}
\usage{
rmniigz(x)
rmnii(x)
rmgz(x)
}
\arguments{
  \item{x}{is the file name.}
}
\value{
  The file name without offending suffix.
}
%\details{}
%\references{}
%\seealso{}
%\examples{}
\author{Brandon Whitcher \email{bjw34032@users.sourceforge.net}}
\keyword{Misc}
