\name{rmniigz}
\alias{rmniigz}
\alias{rmnii}
\alias{rmgz}
\alias{rmhdrgz}
\alias{rmhdr}
\alias{rmimggz}
\alias{rmimg}
\title{Remove File Extensions Around the NIfTI/ANALYZE Formats}
\description{
  Simple function(s) that remove file extensions commonly found when
  using NIfTI-1 or ANALYZE format files.
}
\usage{
rmniigz(x)
rmnii(x)
rmgz(x)
rmhdrgz(x)
rmhdr(x)
rmimggz(x)
rmimg(x)
}
\arguments{
  \item{x}{is the file name.}
}
\value{
  The file name without offending suffix.
}
%\details{}
%\references{}
%\seealso{}
%\examples{}
\author{Brandon Whitcher \email{bwhitcher@gmail.com}}
\keyword{Misc}
