% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/reoriented.R
\docType{methods}
\name{reoriented-methods}
\alias{reoriented}
\alias{reoriented,anlz-method}
\alias{reoriented,nifti-method}
\alias{reoriented-methods}
\alias{reoriented-methods,}
\alias{reoriented<-}
\alias{reoriented<-,anlz-method}
\alias{reoriented<-,nifti-method}
\title{Extract Image Attribute \code{reoriented}}
\usage{
reoriented(object)

\S4method{reoriented}{nifti}(object)

\S4method{reoriented}{anlz}(object)

reoriented(object) <- value

\S4method{reoriented}{nifti}(object) <- value

\S4method{reoriented}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}.}

\item{value}{is the value to assign to the \code{reoriented} field.}
}
\description{
Methods that act on the \code{reoriented} field in the
NIfTI/ANALYZE header.
}
\details{
See documentation on the ANALYZE and/or NIfTI data standards for
more details.
}
\author{
John Muschelli \email{muschellij2@gmail.com},\cr
Brandon Whitcher \email{bwhitcher@gmail.com}
}
\references{
ANALYZE 7.5\cr
\url{http://www.mayo.edu/bir/PDF/ANALYZE75.pdf}\cr
NIfTI-1\cr
\url{http://nifti.nimh.nih.gov/}
}

