% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scl_inter.R
\docType{methods}
\name{scl_inter-methods}
\alias{scl.inter}
\alias{scl.inter,nifti-method}
\alias{scl.inter<-}
\alias{scl.inter<-,nifti-method}
\alias{scl_inter}
\alias{scl_inter,nifti-method}
\alias{scl_inter-methods}
\alias{scl_inter-methods,}
\alias{scl_inter<-}
\alias{scl_inter<-,nifti-method}
\title{Extract Image Attribute \code{scl_inter}}
\usage{
scl_inter(object)

\S4method{scl_inter}{nifti}(object)

scl_inter(object) <- value

\S4method{scl_inter}{nifti}(object) <- value

scl.inter(object)

\S4method{scl.inter}{nifti}(object)

scl.inter(object) <- value

\S4method{scl.inter}{nifti}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}.}

\item{value}{is the value to assign to the \code{scl_inter} field.}
}
\description{
Methods that act on the \code{scl_inter} field in the
NIfTI/ANALYZE header.
}
\details{
See documentation on the ANALYZE and/or NIfTI data standards for
more details.
}
\author{
John Muschelli \email{muschellij2@gmail.com},\cr
Brandon Whitcher \email{bwhitcher@gmail.com}
}
\references{
ANALYZE 7.5\cr
\url{https://rportal.mayo.edu/bir/ANALYZE75.pdf}\cr
NIfTI-1\cr
\url{http://nifti.nimh.nih.gov/}
}

