% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scl_slope.R, R/zzz_niftiImage.R
\docType{methods}
\name{scl_slope-methods}
\alias{scl_slope-methods}
\alias{scl_slope-methods,}
\alias{scl_slope}
\alias{scl_slope,nifti-method}
\alias{scl_slope<-}
\alias{scl_slope<-,nifti-method}
\alias{scl.slope}
\alias{scl.slope,nifti-method}
\alias{scl.slope<-}
\alias{scl.slope<-,nifti-method}
\alias{scl_slope,niftiImage-method}
\alias{scl.slope,niftiImage-method}
\title{Extract Image Attribute \code{scl_slope}}
\usage{
scl_slope(object)

\S4method{scl_slope}{nifti}(object)

scl_slope(object) <- value

\S4method{scl_slope}{nifti}(object) <- value

scl.slope(object)

\S4method{scl.slope}{nifti}(object)

scl.slope(object) <- value

\S4method{scl.slope}{nifti}(object) <- value

\S4method{scl_slope}{niftiImage}(object)

\S4method{scl.slope}{niftiImage}(object)
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}.}

\item{value}{is the value to assign to the \code{scl_slope} field.}
}
\description{
Methods that act on the \code{scl_slope} field in the
NIfTI/ANALYZE header.
}
\details{
See documentation on the ANALYZE and/or NIfTI data standards for
more details.
}
\examples{
file = system.file("extdata", "example.nii.gz", package = "RNifti")
img = RNifti::readNifti(file)
scl_slope(img)
scl.slope(img)
}
\references{
ANALYZE 7.5\cr
\url{http://eeg.sourceforge.net/ANALYZE75.pdf}\cr
NIfTI-1\cr
\url{http://nifti.nimh.nih.gov/}
}
\author{
John Muschelli \email{muschellij2@gmail.com},\cr
Brandon Whitcher \email{bwhitcher@gmail.com}
}
