% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_error.R
\docType{methods}
\name{session_error-methods}
\alias{session_error-methods}
\alias{session_error-methods,}
\alias{session_error}
\alias{session_error,nifti-method}
\alias{session_error,anlz-method}
\alias{session_error<-}
\alias{session_error<-,nifti-method}
\alias{session_error<-,anlz-method}
\alias{session.error}
\alias{session.error,nifti-method}
\alias{session.error,anlz-method}
\alias{session.error<-}
\alias{session.error<-,nifti-method}
\alias{session.error<-,anlz-method}
\title{Extract Image Attribute \code{session_error}}
\usage{
session_error(object)

\S4method{session_error}{nifti}(object)

\S4method{session_error}{anlz}(object)

session_error(object) <- value

\S4method{session_error}{nifti}(object) <- value

\S4method{session_error}{anlz}(object) <- value

session.error(object)

\S4method{session.error}{nifti}(object)

\S4method{session.error}{anlz}(object)

session.error(object) <- value

\S4method{session.error}{nifti}(object) <- value

\S4method{session.error}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}.}

\item{value}{is the value to assign to the \code{session_error} field.}
}
\description{
Methods that act on the \code{session_error} field in the
NIfTI/ANALYZE header.
}
\details{
See documentation on the ANALYZE and/or NIfTI data standards for
more details.
}
\references{
ANALYZE 7.5\cr
\url{http://eeg.sourceforge.net/ANALYZE75.pdf}\cr
NIfTI-1\cr
\url{http://nifti.nimh.nih.gov/}
}
\author{
John Muschelli \email{muschellij2@gmail.com},\cr
Brandon Whitcher \email{bwhitcher@gmail.com}
}
