% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sform_code.R, R/zzz_niftiImage.R
\docType{methods}
\name{sform_code-methods}
\alias{sform_code-methods}
\alias{sform_code-methods,}
\alias{sform_code}
\alias{sform_code,nifti-method}
\alias{sform_code<-}
\alias{sform_code<-,nifti-method}
\alias{sform.code}
\alias{sform.code,nifti-method}
\alias{sform.code<-}
\alias{sform.code<-,nifti-method}
\alias{sform_code,niftiImage-method}
\alias{sform_code<-,niftiImage-method}
\title{Extract Image Attribute \code{sform_code}}
\usage{
sform_code(object)

\S4method{sform_code}{nifti}(object)

sform_code(object) <- value

\S4method{sform_code}{nifti}(object) <- value

sform.code(object)

\S4method{sform.code}{nifti}(object)

sform.code(object) <- value

\S4method{sform.code}{nifti}(object) <- value

\S4method{sform_code}{niftiImage}(object)

\S4method{sform_code}{niftiImage}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}.}

\item{value}{is the value to assign to the \code{sform_code} field.}
}
\description{
Methods that act on the \code{sform_code} field in the
NIfTI/ANALYZE header.
}
\details{
See documentation on the ANALYZE and/or NIfTI data standards for
more details.
}
\examples{
file = system.file("extdata", "example.nii.gz", package = "RNifti")
img = RNifti::readNifti(file)
sform_code(img)
sform_code(img) = 4
stopifnot(sform_code(img) ==  4)
}
\references{
ANALYZE 7.5\cr
\url{http://eeg.sourceforge.net/ANALYZE75.pdf}\cr
NIfTI-1\cr
\url{http://nifti.nimh.nih.gov/}
}
\author{
John Muschelli \email{muschellij2@gmail.com},\cr
Brandon Whitcher \email{bwhitcher@gmail.com}
}
