% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizeof_hdr.R
\docType{methods}
\name{sizeof_hdr-methods}
\alias{sizeof_hdr-methods}
\alias{sizeof_hdr-methods,}
\alias{sizeof_hdr}
\alias{sizeof_hdr,nifti-method}
\alias{sizeof_hdr,anlz-method}
\alias{sizeof.hdr}
\alias{sizeof.hdr,nifti-method}
\alias{sizeof.hdr,anlz-method}
\title{Extract Image Attribute \code{sizeof_hdr}}
\usage{
sizeof_hdr(object)

\S4method{sizeof_hdr}{nifti}(object)

\S4method{sizeof_hdr}{anlz}(object)

sizeof.hdr(object)

\S4method{sizeof.hdr}{nifti}(object)

\S4method{sizeof.hdr}{anlz}(object)
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}.}
}
\description{
Methods that act on the \code{sizeof_hdr} field in the
NIfTI/ANALYZE header.
}
\details{
See documentation on the ANALYZE and/or NIfTI data standards for
more details.
}
\references{
ANALYZE 7.5\cr
\url{https://rportal.mayo.edu/bir/ANALYZE75.pdf}\cr
NIfTI-1\cr
\url{http://nifti.nimh.nih.gov/}
}
\author{
John Muschelli \email{muschellij2@gmail.com},\cr
Brandon Whitcher \email{bwhitcher@gmail.com}
}
