% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/slice_code.R
\docType{methods}
\name{slice_code-methods}
\alias{slice_code}
\alias{slice_code,anlz-method}
\alias{slice_code,nifti-method}
\alias{slice_code-methods}
\alias{slice_code<-}
\alias{slice_code<-,anlz-method}
\alias{slice_code<-,nifti-method}
\title{Extract Image slice_code attribute}
\usage{
slice_code(object)

\S4method{slice_code}{nifti}(object)

\S4method{slice_code}{anlz}(object)

slice_code(object) <- value

\S4method{slice_code}{nifti}(object) <- value

\S4method{slice_code}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to slice_code}
}
\description{
Methods that act on the ``slice_code'' in the NIfTI/ANALYZE header.
}

