% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_duration.R, R/zzz_niftiImage.R
\docType{methods}
\name{slice_duration-methods}
\alias{slice_duration-methods}
\alias{slice_duration-methods,}
\alias{slice_duration}
\alias{slice_duration,nifti-method}
\alias{slice_duration<-}
\alias{slice_duration<-,nifti-method}
\alias{slice.duration}
\alias{slice.duration,nifti-method}
\alias{slice.duration<-}
\alias{slice.duration<-,nifti-method}
\alias{slice_duration,niftiImage-method}
\alias{slice_duration<-,niftiImage-method}
\title{Extract Image Attribute \code{slice_duration}}
\usage{
slice_duration(object)

\S4method{slice_duration}{nifti}(object)

slice_duration(object) <- value

\S4method{slice_duration}{nifti}(object) <- value

slice.duration(object)

\S4method{slice.duration}{nifti}(object)

slice.duration(object) <- value

\S4method{slice.duration}{nifti}(object) <- value

\S4method{slice_duration}{niftiImage}(object)

\S4method{slice_duration}{niftiImage}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}.}

\item{value}{is the value to assign to the \code{slice_duration} field.}
}
\description{
Methods that act on the \code{slice_duration} field in the
NIfTI/ANALYZE header.
}
\details{
See documentation on the ANALYZE and/or NIfTI data standards for
more details.
}
\examples{
file = system.file("extdata", "example.nii.gz", package = "RNifti")
img = RNifti::readNifti(file)
slice_duration(img)
slice_duration(img) = 8
stopifnot(slice_duration(img) ==  8)
}
\references{
ANALYZE 7.5\cr
\url{http://eeg.sourceforge.net/ANALYZE75.pdf}\cr
NIfTI-1\cr
\url{http://nifti.nimh.nih.gov/}
}
\author{
John Muschelli \email{muschellij2@gmail.com},\cr
Brandon Whitcher \email{bwhitcher@gmail.com}
}
