% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/slice_start.R
\docType{methods}
\name{slice_start}
\alias{slice_start}
\alias{slice_start,anlz-method}
\alias{slice_start,nifti-method}
\alias{slice_start-methods}
\alias{slice_start<-}
\alias{slice_start<-,anlz-method}
\alias{slice_start<-,nifti-method}
\title{Extract NIfTI 3D Image slice_start attribute}
\usage{
slice_start(object)

\S4method{slice_start}{nifti}(object)

\S4method{slice_start}{anlz}(object)

slice_start(object) <- value

\S4method{slice_start}{nifti}(object) <- value

\S4method{slice_start}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti}}

\item{value}{Value to assign to slice_start}
}
\description{
Methods that act on the ``slice_start'' in the NIfTI header.
}

