% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_start.R, R/zzz_niftiImage.R
\docType{methods}
\name{slice_start-methods}
\alias{slice_start-methods}
\alias{slice_start-methods,}
\alias{slice_start}
\alias{slice_start,nifti-method}
\alias{slice_start<-}
\alias{slice_start<-,nifti-method}
\alias{slice.start}
\alias{slice.start,nifti-method}
\alias{slice.start<-}
\alias{slice.start<-,nifti-method}
\alias{slice_start,niftiImage-method}
\alias{slice_start<-,niftiImage-method}
\alias{slice.start,niftiImage-method}
\alias{slice.start<-,niftiImage-method}
\title{Extract Image Attribute \code{slice_start}}
\usage{
slice_start(object)

\S4method{slice_start}{nifti}(object)

slice_start(object) <- value

\S4method{slice_start}{nifti}(object) <- value

slice.start(object)

\S4method{slice.start}{nifti}(object)

slice.start(object) <- value

\S4method{slice.start}{nifti}(object) <- value

\S4method{slice_start}{niftiImage}(object)

\S4method{slice_start}{niftiImage}(object) <- value

\S4method{slice.start}{niftiImage}(object)

\S4method{slice.start}{niftiImage}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}.}

\item{value}{is the value to assign to the \code{slice_start} field.}
}
\description{
Methods that act on the \code{slice_start} field in the
NIfTI/ANALYZE header.
}
\details{
See documentation on the ANALYZE and/or NIfTI data standards for
more details.
}
\examples{
file = system.file("extdata", "example.nii.gz", package = "RNifti")
img = RNifti::readNifti(file)
slice_start(img)
slice_start(img) = 4
stopifnot(slice_start(img) == 4) 
slice.start(img)
slice.start(img) = 0
}
\references{
ANALYZE 7.5\cr
\url{http://eeg.sourceforge.net/ANALYZE75.pdf}\cr
NIfTI-1\cr
\url{http://nifti.nimh.nih.gov/}
}
\author{
John Muschelli \email{muschellij2@gmail.com},\cr
Brandon Whitcher \email{bwhitcher@gmail.com}
}
