% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srow_y.R, R/zzz_niftiImage.R
\docType{methods}
\name{srow_y-methods}
\alias{srow_y-methods}
\alias{srow_y-methods,}
\alias{srow_y}
\alias{srow_y,nifti-method}
\alias{srow_y<-}
\alias{srow_y<-,nifti-method}
\alias{srow.y}
\alias{srow.y,nifti-method}
\alias{srow.y<-}
\alias{srow.y<-,nifti-method}
\alias{srow_y,niftiImage-method}
\alias{srow_y<-,niftiImage-method}
\title{Extract Image Attribute \code{srow_y}}
\usage{
srow_y(object)

\S4method{srow_y}{nifti}(object)

srow_y(object) <- value

\S4method{srow_y}{nifti}(object) <- value

srow.y(object)

\S4method{srow.y}{nifti}(object)

srow.y(object) <- value

\S4method{srow.y}{nifti}(object) <- value

\S4method{srow_y}{niftiImage}(object)

\S4method{srow_y}{niftiImage}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}.}

\item{value}{is the value to assign to the \code{srow_y} field.}
}
\description{
Methods that act on the \code{srow_y} field in the
NIfTI/ANALYZE header.
}
\details{
See documentation on the ANALYZE and/or NIfTI data standards for
more details.
}
\references{
ANALYZE 7.5\cr
\url{http://eeg.sourceforge.net/ANALYZE75.pdf}\cr
NIfTI-1\cr
\url{http://nifti.nimh.nih.gov/}
}
\author{
John Muschelli \email{muschellij2@gmail.com},\cr
Brandon Whitcher \email{bwhitcher@gmail.com}
}
