% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srow_z.R, R/zzz_niftiImage.R
\docType{methods}
\name{srow_z-methods}
\alias{srow_z-methods}
\alias{srow_z-methods,}
\alias{srow_z}
\alias{srow_z,nifti-method}
\alias{srow_z<-}
\alias{srow_z<-,nifti-method}
\alias{srow.z}
\alias{srow.z,nifti-method}
\alias{srow.z<-}
\alias{srow.z<-,nifti-method}
\alias{srow_z,niftiImage-method}
\alias{srow_z<-,niftiImage-method}
\title{Extract Image Attribute \code{srow_z}}
\usage{
srow_z(object)

\S4method{srow_z}{nifti}(object)

srow_z(object) <- value

\S4method{srow_z}{nifti}(object) <- value

srow.z(object)

\S4method{srow.z}{nifti}(object)

srow.z(object) <- value

\S4method{srow.z}{nifti}(object) <- value

\S4method{srow_z}{niftiImage}(object)

\S4method{srow_z}{niftiImage}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}.}

\item{value}{is the value to assign to the \code{srow_z} field.}
}
\description{
Methods that act on the \code{srow_z} field in the
NIfTI/ANALYZE header.
}
\details{
See documentation on the ANALYZE and/or NIfTI data standards for
more details.
}
\references{
ANALYZE 7.5\cr
\url{http://eeg.sourceforge.net/ANALYZE75.pdf}\cr
NIfTI-1\cr
\url{http://nifti.nimh.nih.gov/}
}
\author{
John Muschelli \email{muschellij2@gmail.com},\cr
Brandon Whitcher \email{bwhitcher@gmail.com}
}
