% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toffset.R, R/zzz_niftiImage.R
\docType{methods}
\name{toffset-methods}
\alias{toffset-methods}
\alias{toffset-methods,}
\alias{toffset}
\alias{toffset,nifti-method}
\alias{toffset<-}
\alias{toffset<-,nifti-method}
\alias{toffset,niftiImage-method}
\alias{toffset<-,niftiImage-method}
\title{Extract Image Attribute \code{toffset}}
\usage{
toffset(object)

\S4method{toffset}{nifti}(object)

toffset(object) <- value

\S4method{toffset}{nifti}(object) <- value

\S4method{toffset}{niftiImage}(object)

\S4method{toffset}{niftiImage}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}.}

\item{value}{is the value to assign to the \code{toffset} field.}
}
\description{
Methods that act on the \code{toffset} field in the
NIfTI/ANALYZE header.
}
\details{
See documentation on the ANALYZE and/or NIfTI data standards for
more details.
}
\examples{
file = system.file("extdata", "example.nii.gz", package = "RNifti")
img = RNifti::readNifti(file)
toffset(img)
toffset(img) = 8
stopifnot(toffset(img) ==  8)
}
\references{
ANALYZE 7.5\cr
\url{http://eeg.sourceforge.net/ANALYZE75.pdf}\cr
NIfTI-1\cr
\url{http://nifti.nimh.nih.gov/}
}
\author{
John Muschelli \email{muschellij2@gmail.com},\cr
Brandon Whitcher \email{bwhitcher@gmail.com}
}
