% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unused1.R
\docType{methods}
\name{unused1-methods}
\alias{unused1-methods}
\alias{unused1-methods,}
\alias{unused1}
\alias{unused1,anlz-method}
\alias{unused1<-}
\alias{unused1<-,anlz-method}
\title{Extract Image Attribute \code{unused1}}
\usage{
unused1(object)

\S4method{unused1}{anlz}(object)

unused1(object) <- value

\S4method{unused1}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}.}

\item{value}{is the value to assign to the \code{unused1} field.}
}
\description{
Methods that act on the \code{unused1} field in the
NIfTI/ANALYZE header.
}
\details{
See documentation on the ANALYZE and/or NIfTI data standards for
more details.
}
\references{
ANALYZE 7.5\cr
\url{http://eeg.sourceforge.net/ANALYZE75.pdf}\cr
NIfTI-1\cr
\url{http://nifti.nimh.nih.gov/}
}
\author{
John Muschelli \email{muschellij2@gmail.com},\cr
Brandon Whitcher \email{bwhitcher@gmail.com}
}
