% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vox_offset.R, R/zzz_niftiImage.R
\docType{methods}
\name{vox_offset-methods}
\alias{vox_offset-methods}
\alias{vox_offset-methods,}
\alias{vox_offset}
\alias{vox_offset,nifti-method}
\alias{vox_offset,anlz-method}
\alias{vox_offset<-}
\alias{vox_offset<-,nifti-method}
\alias{vox_offset<-,anlz-method}
\alias{vox.offset}
\alias{vox.offset,nifti-method}
\alias{vox.offset,anlz-method}
\alias{vox.offset<-}
\alias{vox.offset<-,nifti-method}
\alias{vox.offset<-,anlz-method}
\alias{vox_offset,niftiImage-method}
\alias{vox.offset,niftiImage-method}
\title{Extract Image Attribute \code{vox_offset}}
\usage{
vox_offset(object)

\S4method{vox_offset}{nifti}(object)

\S4method{vox_offset}{anlz}(object)

vox_offset(object) <- value

\S4method{vox_offset}{nifti}(object) <- value

\S4method{vox_offset}{anlz}(object) <- value

vox.offset(object)

\S4method{vox.offset}{nifti}(object)

\S4method{vox.offset}{anlz}(object)

vox.offset(object) <- value

\S4method{vox.offset}{nifti}(object) <- value

\S4method{vox.offset}{anlz}(object) <- value

\S4method{vox_offset}{niftiImage}(object)

\S4method{vox.offset}{niftiImage}(object)
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}.}

\item{value}{is the value to assign to the \code{vox_offset} field.}
}
\description{
Methods that act on the \code{vox_offset} field in the
NIfTI/ANALYZE header.
}
\details{
See documentation on the ANALYZE and/or NIfTI data standards for
more details.
}
\examples{
file = system.file("extdata", "example.nii.gz", package = "RNifti")
img = RNifti::readNifti(file)
vox_offset(img)
img = RNifti::readNifti(file)
vox.offset(img)
}
\references{
ANALYZE 7.5\cr
\url{http://eeg.sourceforge.net/ANALYZE75.pdf}\cr
NIfTI-1\cr
\url{http://nifti.nimh.nih.gov/}
}
\author{
John Muschelli \email{muschellij2@gmail.com},\cr
Brandon Whitcher \email{bwhitcher@gmail.com}
}
