% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vox_units.R
\docType{methods}
\name{vox_units-methods}
\alias{vox_units-methods}
\alias{vox_units-methods,}
\alias{vox_units}
\alias{vox_units,anlz-method}
\alias{vox_units<-}
\alias{vox_units<-,anlz-method}
\alias{vox.units}
\alias{vox.units,nifti-method}
\alias{vox.units,anlz-method}
\alias{vox.units<-}
\alias{vox.units<-,anlz-method}
\title{Extract Image Attribute \code{vox_units}}
\usage{
vox_units(object)

\S4method{vox_units}{anlz}(object)

vox_units(object) <- value

\S4method{vox_units}{anlz}(object) <- value

vox.units(object)

\S4method{vox.units}{anlz}(object)

vox.units(object) <- value

\S4method{vox.units}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}.}

\item{value}{is the value to assign to the \code{vox_units} field.}
}
\description{
Methods that act on the \code{vox_units} field in the
NIfTI/ANALYZE header.
}
\details{
See documentation on the ANALYZE and/or NIfTI data standards for
more details.
}
\references{
ANALYZE 7.5\cr
\url{http://eeg.sourceforge.net/ANALYZE75.pdf}\cr
NIfTI-1\cr
\url{http://nifti.nimh.nih.gov/}
}
\author{
John Muschelli \email{muschellij2@gmail.com},\cr
Brandon Whitcher \email{bwhitcher@gmail.com}
}
