% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voxdim.R
\name{voxdim}
\alias{voxdim}
\title{Gets Voxel Dimensions}
\usage{
voxdim(img)
}
\arguments{
\item{img}{nifti object}
}
\value{
Vector of length 3
}
\description{
Grabs the pixdim and takes the correct elements
}
\examples{
nim <- nifti(array(rnorm(10^3), dim = c(5, 2, 100)),
pixdim = c(1, 0.5, 0.2, 1))
voxdim(nim)

}
