% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xyzt_units.R
\docType{methods}
\name{xyzt_units-methods}
\alias{xyzt_units}
\alias{xyzt_units,anlz-method}
\alias{xyzt_units,nifti-method}
\alias{xyzt_units-methods}
\alias{xyzt_units<-}
\alias{xyzt_units<-,anlz-method}
\alias{xyzt_units<-,nifti-method}
\title{Extract Image xyzt_units attribute}
\usage{
xyzt_units(object)

\S4method{xyzt_units}{nifti}(object)

\S4method{xyzt_units}{anlz}(object)

xyzt_units(object) <- value

\S4method{xyzt_units}{nifti}(object) <- value

\S4method{xyzt_units}{anlz}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{nifti} or \code{anlz}}

\item{value}{Value to assign to xyzt_units}
}
\description{
Methods that act on the ``xyzt_units'' in the NIfTI/ANALYZE header.
}

